% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_input_data.R
\name{append_lead_trail_na}
\alias{append_lead_trail_na}
\title{Append Leading and Trailing NA's}
\usage{
append_lead_trail_na(df, na)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{na}{\code{data.frame} containing rows with leading and/or trailing
\code{NA}.}
}
\description{
\code{append_lead_trail_na} appends leading and trailing \code{NA} back
  to the \code{data.frame}.
}
\keyword{internal}
