% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{axis_labels_period}
\alias{axis_labels_period}
\title{Define Axis Labels Based on Period}
\usage{
axis_labels_period(df, plot_period, tz)
}
\arguments{
\item{df}{input \code{data.frame}}

\item{plot_period}{character, specify whether plots should be displayed for
the whole period (\code{plot_period = "full"}), separately for each
year (\code{plot_period = "yearly"}) or for each month
(\code{plot_period = "monthly"}).}

\item{tz}{specify the desired time zone.}
}
\description{
\code{axis_labels_period} defines axis ticks and labels based on the
  period used for plotting.
}
\keyword{internal}
