% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{calccycle}
\alias{calccycle}
\title{Calculates Cycle Statistics}
\usage{
calccycle(df, reso, tz, plot_cycle = FALSE, plot_export)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}

\item{plot_cycle}{logical, specify whether maxima and minima used for the
calculation of the cycle statistics should be plotted.}

\item{plot_export}{logical, specifies whether the plots are exported as a
\code{PDF} file to the working directory or are plotted in the console.}
}
\description{
\code{calccycle} calculates different statistics and characteristics of
  cycles. To identify the cycles local maxima and minima are identified
  using a moving window.
}
\details{
\code{calccycle} is inspired by the function
  \code{\link[dendrometeR]{phase_def}} in the package \code{dendrometeR}.
}
\keyword{internal}
