% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_cycle.R
\name{calccycleparam}
\alias{calccycleparam}
\title{Calculate Cycle Parameters}
\usage{
calccycleparam(df, maxmin, mode)
}
\arguments{
\item{df}{\code{data.frame} with dendrometer data}

\item{maxmin}{\code{data.frame} with minima and maxima and the respective
time stamps.}

\item{mode}{character, specify whether statistics for shrinkage
(\code{mode = "shrink"}) or refilling \code{mode = "ref"} are
calculated.}
}
\description{
\code{calccycleparam} calculates parameters for cycles of shrinking and
  refilling.
}
\keyword{internal}
