% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{calcdiff}
\alias{calcdiff}
\title{Calculates Hourly Change in Data}
\usage{
calcdiff(df, reso)
}
\arguments{
\item{df}{input \code{data.frame}.}
}
\description{
\code{calcdiff} calculates an hourly change in \code{value}. In case there
  are gaps, the difference is also calculated at an hourly rate respective
  to the length of the preceding gap.
}
\examples{

}
\keyword{internal}
