% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{calcgroperiods}
\alias{calcgroperiods}
\title{Calculate Growth for Different Time Periods}
\usage{
calcgroperiods(df, reso, tz)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{reso}{numeric, desired output time resolution (in minutes). See
Details for more information on data aggregation.}

\item{tz}{specify the desired time zone.}
}
\description{
\code{calcgroperiods} calculates the minimum, median and maximum growth
  for different time periods. The periods are selected depending on the
  resolution. Growth values are calculated after removing periods without
  growth (i.e. growth = 0).
}
\keyword{internal}
