% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{calcmax}
\alias{calcmax}
\title{Calculate Maximum}
\usage{
calcmax(df)
}
\arguments{
\item{df}{input \code{data.frame}.}
}
\description{
\code{calcmax} calculates the maximum of measured values.
}
\keyword{internal}
