% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{calcmds}
\alias{calcmds}
\title{Calculate Maximum Daily Shrinkage (MDS)}
\usage{
calcmds(df, reso, tz, plot_mds = FALSE, plot_export)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}

\item{plot_mds}{logical, specify whether maxima and minima used for the
calculation of the maximum daily shrinkage (\code{mds}) should be
plotted.}

\item{plot_export}{logical, specifies whether the plots are exported as a
\code{PDF} file to the working directory or are plotted in the console.}
}
\description{
\code{calcmds} calculates the maximum daily shrinkage (\code{mds}).
  \code{mds} is defined as the largest daily shrinkage. First, local
  maxima and minima are identified using a moving window. \code{mds} is
  only calculated if a local maximum occurs before a local minimum
  (i.e. if the stem shrinks during the day).
}
\details{
\code{calcmds} is inspired by the function
  \code{\link[dendrometeR]{phase_def}} in the package \code{dendrometeR}.
}
\keyword{internal}
