% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{calcmissing}
\alias{calcmissing}
\title{Calculates Percentages of Interpolated, Deleted and Missing Data}
\usage{
calcmissing(data_plot)
}
\arguments{
\item{data_plot}{input \code{data.frame} containing both \code{L1} and
\code{L2} data as well as changes in \code{L2} for plotting.}
}
\value{
a list of length = 3 with percentages of interpolated, deleted,
  and missing data.
}
\description{
\code{calcmissing} calculates the percentage of interpolated, deleted and
  missing data.
}
\keyword{internal}
