% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{calctwdgro}
\alias{calctwdgro}
\title{Calculates TWD and GRO}
\usage{
calctwdgro(df, tz)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{tz}{specify the desired time zone. Default is \code{"Etc/GMT-1"}.}
}
\description{
\code{calctwdgro} calculates the tree water deficit (twd) and the growth
  since the beginning of the year (gro_yr).
}
\examples{

}
\keyword{internal}
