% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_variables.R
\name{check_data_L1}
\alias{check_data_L1}
\title{Check data_L1 Input}
\usage{
check_data_L1(data_L1)
}
\arguments{
\item{data_L1}{time-aligned dendrometer data as produced by
\code{\link{proc_L1}}.}
}
\description{
\code{check_data_L1} checks the input data given to the variable
  \code{\link{data_L1}}.
}
\examples{

}
\keyword{internal}
