% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_variables.R
\name{check_data_L2}
\alias{check_data_L2}
\title{Check data_L2 Input}
\usage{
check_data_L2(data_L2)
}
\arguments{
\item{data_L2}{processed dendrometer data as produced by
\code{\link{proc_dendro_L2}}.}
}
\description{
\code{check_data_L2} checks the input data given to the variable
  \code{\link{data_L1}}.
}
\examples{

}
\keyword{internal}
