% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_variables.R
\name{check_date_period}
\alias{check_date_period}
\title{Check Date Input for Period}
\usage{
check_date_period(datevec, datevec_name, df)
}
\arguments{
\item{datevec}{character vector in a standard date or datetime format.}

\item{df}{input \code{data.frame}.}
}
\description{
\code{check_date_period} checks whether the provided dates overlap with
  the period of the data.
}
\keyword{internal}
