% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_variables.R
\name{check_delete}
\alias{check_delete}
\title{Check the Date Sequence in Delete}
\usage{
check_delete(delete)
}
\description{
\code{check_delete} checks whether the first value in a pair of dates
  occurs before the second value.
}
\keyword{internal}
