% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_input_data.R
\name{check_format}
\alias{check_format}
\title{Check Format of Input Data}
\usage{
check_format(df, input)
}
\arguments{
\item{df}{input \code{data.frame}.}
}
\description{
\code{check_format()} checks whether input table is in long or wide format.
}
\keyword{internal}
