% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_variables.R
\name{check_input_variables}
\alias{check_input_variables}
\title{Check all input variables automatically}
\usage{
check_input_variables(list)
}
\arguments{
\item{list}{list of all input arguments to be checked.}
}
\description{
\code{check_input_variables} checks all input variables of exported
  functions.
}
\keyword{internal}
