% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_variables.R
\name{check_logical}
\alias{check_logical}
\title{Check for Logical Input}
\usage{
check_logical(var, var_name)
}
\arguments{
\item{var}{variable to be checked.}

\item{var_name}{character, name of variable to be checked.}
}
\description{
\code{check_logical} checks whether input variable is of type logical.
}
\keyword{internal}
