% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_data.R
\name{check_missing}
\alias{check_missing}
\title{Check for Sensors with Missing Data}
\usage{
check_missing(df)
}
\arguments{
\item{df}{input \code{data.frame}.}
}
\description{
\code{check_missing} removes series without any data or with the same
  value over the whole period.
}
\keyword{internal}
