% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_variables.R
\name{check_numeric}
\alias{check_numeric}
\title{Check for Numeric Input}
\usage{
check_numeric(var, var_name)
}
\arguments{
\item{var}{variable to be checked.}

\item{var_name}{character, name of variable to be checked.}
}
\description{
\code{check_numeric} checks whether input variable is of type numeric.
}
\keyword{internal}
