% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_variables.R
\name{check_package}
\alias{check_package}
\title{Check if Package is Installed}
\usage{
check_package(pck_name)
}
\arguments{
\item{pck_name}{character, name of the required package.}
}
\description{
\code{check_package} checks whether a suggested package that is needed
  for a specific function is already installed.
}
\keyword{internal}
