% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_variables.R
\name{check_series}
\alias{check_series}
\title{Checks Input of Series}
\usage{
check_series(df, series)
}
\arguments{
\item{df}{input \code{data.frame}.}
}
\description{
\code{check_series} checks the input of the variable series.
}
\keyword{internal}
