% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_data.R
\name{check_ts}
\alias{check_ts}
\title{Check TS Column}
\usage{
check_ts(df, date_format, tz)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{date_format}{character, specify a custom date format if it is not
\code{"\%Y-\%m-\%d \%H:\%M:\%S"}.}

\item{tz}{specify the desired time zone.}
}
\description{
\code{check_ts} checks whether the 'ts' column is in a standard date
  or datetime format.
}
\keyword{internal}
