% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{cleanoutofrange}
\alias{cleanoutofrange}
\title{Process Dendrometer Data}
\usage{
cleanoutofrange(df, val_range)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{val_range}{numeric vector specifying the minimum and maximum
\code{c(min, max)} of credible dendrometer measurement values. Values
lower than \code{min} or higher than \code{max} are deleted without
notice.}
}
\description{
\code{cleanoutofrange} removes implausible data points lower or higher
  than specified values in \code{val_range}.
}
\examples{

}
\keyword{internal}
