% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_dendro_L1.R
\name{corr_dendro_L1}
\alias{corr_dendro_L1}
\title{Manually Correct Time-Aligned Dendrometer Data}
\usage{
corr_dendro_L1(
  data_L1,
  delete,
  series = NULL,
  plot = TRUE,
  plot_export = TRUE,
  tz = "UTC"
)
}
\arguments{
\item{data_L1}{time-aligned dendrometer data as produced by
\code{\link{proc_L1}}.}

\item{plot}{logical, plots the corrected dataset with the original data
in the background in grey.}

\item{plot_export}{logical, specifies whether the plots are exported as a
\code{PDF} file to the working directory or are plotted in the console.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\value{
The function returns a \code{data.frame} with corrected \code{L1}
  dendrometer data.
}
\description{
\code{corr_dendro_L1} deletes values in periods with erroneous data of
  time-aligned (\code{L1}) dendrometer or climate data. Values are
  overwritten with \code{NA} without further notice.
}
\examples{
corr_dendro_L1(data_L1 = dendro_data_L1,
               delete = c("2013-08-01", "2013-08-05"),
               series = "site-1_dendro-3", plot_export = FALSE)

}
\seealso{
\code{\link{corr_dendro_L3}} to correct \code{L2} data.
}
