% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_dendro_L2.R
\name{corr_dendro_L2}
\alias{corr_dendro_L2}
\title{Manually Correct Processed Dendrometer Data}
\usage{
corr_dendro_L2(
  dendro_L1 = NULL,
  dendro_L2,
  reverse = NULL,
  force = NULL,
  delete = NULL,
  series = NULL,
  n_days = 5,
  plot = TRUE,
  plot_export = TRUE,
  plot_name = "corr_L2_plot",
  tz = "UTC"
)
}
\arguments{
\item{dendro_L1}{time-aligned dendrometer data as produced by
\code{\link{proc_L1}}. Optional, only needed for \code{reverse} and if
\code{plot = TRUE}.}

\item{dendro_L2}{cleaned dendrometer data as produced by
\code{\link{proc_dendro_L2}}.}

\item{reverse}{numeric vector, specify ID numbers of the changes that should
be reversed. ID numbers are reported in the plots produced by
\code{\link{proc_dendro_L2}} or \code{\link{plot_proc_L2}} with the
argument \code{plot_period = "monthly"}.}

\item{force}{character vector, specify the dates after which jumps/shifts
should be corrected. The largest value difference occurring in a
period of \code{n_days} after the specified dates in \code{force} is
corrected. Dates need to be in a standard date or
datetime format (e.g. \code{\%Y-\%m-\%d \%H:\%M:\%S}).}

\item{delete}{character vector, specify pairs of dates between which
all dendrometer data will be deleted (i.e. 4 dates will result in two
periods: 1-2 and 3-4 in which data is deleted). Dates need to be in the
same standard date or datetime format
(e.g. \code{\%Y-\%m-\%d \%H:\%M:\%S}).}

\item{series}{character, specify the name of a single dendrometer series
for which changes should be made. Data of other series is left unchanged.
Not needed if only a single series is provided.}

\item{n_days}{numeric, length of the period (in days) following the dates
specified in \code{force} in which a jump/shift is corrected. Increase if
the gap in data is longer than the default (\code{n_days = 5}).}

\item{plot}{logical, specify whether implemented changes should be plotted.}

\item{plot_export}{logical, specifies whether the plots are exported as a
\code{PDF} file to the working directory or are plotted in the console.}

\item{plot_name}{character, specify name of the PDF to which the plots are
saved.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\value{
The function returns a \code{data.frame} with corrected \code{L2}
  dendrometer data. The corrections are documented in the column
  \code{flags}.
}
\description{
\code{corr_dendro_L2} corrects remaining errors in cleaned \code{L2}
  dendrometer data. The function can be used to manually correct remaining
  errors that cannot be removed by adjusting the input parameter values in
  \code{\link{proc_dendro_L2}}. The function can reverse erroneously
  introduced changes or force changes that were not automatically made.
}
\examples{
corr_dendro_L2(dendro_L1 = dendro_data_L1, dendro_L2 = dendro_data_L2,
               reverse = 59:61, force = "2013-08-12",
               delete = c("2013-08-01", "2013-08-04"),
               series = "site-1_dendro-3", plot_export = FALSE,
               plot_name = "corr_L2_plot")

}
\seealso{
\code{\link{corr_dendro_L1}} to correct \code{L1} data.
}
