% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_data.R
\name{create_temp_dummy}
\alias{create_temp_dummy}
\title{Create Dummy Temperature Dataset}
\usage{
create_temp_dummy(df)
}
\arguments{
\item{df}{input \code{data.frame} of dendrometer data.}
}
\description{
\code{create_temp_dummy} creates a dummy temperature dataset if local
  temperature measurements are missing. Temperatures in the months December,
  January and February are 0°C (i.e. frost shrinkage is possible).
  Tempeartures in the other months are 10°C (i.e. no frost shrinkage
  assumed).
}
\keyword{internal}
