% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{createflagdiff}
\alias{createflagdiff}
\title{Create Flag for Abrubt Changes in Data}
\usage{
createflagdiff(df, reso, diffwin = 1000, diffsum = 150)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{diffwin}{maximal hourly difference expected in winter.}

\item{diffsum}{maximal hourly difference expected in summer.}
}
\description{
\code{createflagdiff} creates a flag for large differences in value.
}
\examples{

}
\keyword{internal}
