% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{createflagfragment}
\alias{createflagfragment}
\title{Create Flag for Small Data Fragments}
\usage{
createflagfragment(df, frag_len = NULL)
}
\arguments{
\item{frag_len}{numeric, specifying the length of non-\code{NA} data in
between \code{NA} data. This is often the case if there is erroneous data
in between jumps. If \code{frag_len = NULL} the devault value is set to
\code{frag_len = 2.1}.}

\item{out}{logical, vector with outlier flags.}
}
\description{
\code{createflagfragment} flags short fragments of not flagged data that
  is in-between flagged data. Such cases occur often in spikes of outlier
  data.
}
\keyword{internal}
