% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{createflagmad}
\alias{createflagmad}
\title{Creates Flag for Outliers Based on Median Absolute Deviation}
\usage{
createflagmad(df, reso, wnd = 6, n_mad = 8)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{wnd}{length of time window for which \code{mad} is calculated. A
second longer time window (\code{wnd * 5}) is automatically added. Unit
of \code{wnd} is days.}

\item{n_mad}{number of \code{mad} distances from the
first and third quartile that are still trusted. \code{n_mad} is
increased to \code{n_mad * 10} if there is a probability of frost.}
}
\description{
\code{createflagmad} creates flag for outliers that are above or below a
  threshold distant from the first or third quantile. The threshold is
  specified by the the median absolute deviation
  (\code{\link[stats]{mad}}).
}
\examples{

}
\keyword{internal}
