% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{createfrostflag}
\alias{createfrostflag}
\title{Creates Flag for Potential Frost}
\usage{
createfrostflag(df, tem, lowtemp = 5)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{lowtemp}{numeric, specifies temperature in °C below which shrinkage
in stem diameter due to frost is expected.}
}
\description{
\code{createfrostflag} adds a flag for potential frost.
}
\examples{

}
\keyword{internal}
