% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{creategapflag}
\alias{creategapflag}
\title{Creates Flag for Gaps}
\usage{
creategapflag(df, reso, gaple = 12 * (60/reso))
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{gaple}{specify minimum length of gap for flagging.}
}
\description{
\code{creategapflag} adds a flag to gaps that are longer than
  \code{gaple}. Used to be called \code{cleanaftergaps}.
}
\keyword{internal}
