% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{createjumpflag}
\alias{createjumpflag}
\title{Creates Flag for Jumps in Data}
\usage{
createjumpflag(df)
}
\arguments{
\item{df}{input \code{data.frame}.}
}
\description{
\code{createjumpflag} creates flag for jumps.
}
\keyword{internal}
