% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{createjumpoutflag}
\alias{createjumpoutflag}
\title{Creates Flag for Outliers and Jumps in Data}
\usage{
createjumpoutflag(df, thr = 0.2)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{thr}{specifies the threshold to discriminate between outliers and
jumps due to adjustments of the dendrometer needle.}
}
\description{
\code{createjumpoutflag} creates flag for jumps and ouliers.
}
\examples{

}
\keyword{internal}
