% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycle_stats.R
\name{cycle_stats}
\alias{cycle_stats}
\title{Calculates Cycle Statistics}
\usage{
cycle_stats(df, reso, tz, plot_cycle = FALSE, plot_export)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{reso}{desired time resolution of output (in minutes). See
\code{Details} for more information on data aggregation.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}

\item{plot_cycle}{logical, specify whether maxima and minima used for the
calculation of the cycle statistics should be plotted.}

\item{plot_export}{logical, specifies whether the plots are exported as a
\code{PDF} file to the working directory or are plotted in the console.}
}
\description{
\code{cycle_stats} calculates different statistics and characteristics of
  cycles. To identify the cycles local maxima and minima are identified
  using a moving window.
}
\details{
\code{cycle_stats} is inspired by the function
  \code{\link[dendrometeR]{phase_def}} in the package \code{dendrometeR}.
}
\examples{

}
