% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dendro_data_L0}
\alias{dendro_data_L0}
\title{Raw Dendrometer Data in Long Format}
\format{
A data frame with 152469 rows and 3 variables:
\describe{
  \item{series_id}{name of the data series}
  \item{ts}{timestamp of the measurements}
  \item{value}{measured values}
}
}
\usage{
data(dendro_data_L0)
}
\description{
A dataset containing raw dendrometer data that is formatted in
long format.
}
\keyword{datasets}
