% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dendro_data_L1}
\alias{dendro_data_L1}
\title{Time-aligned Dendrometer Data}
\format{A data frame with 13642 rows and 4 variables.}
\usage{
data(dendro_data_L1)
}
\value{
a \code{data.frame} with measurements aligned to regular time
  intervals (interval specified in \code{reso}) containing the following
  columns:
   \item{series}{name of the series.}
   \item{ts}{timestamp with format \code{\%Y-\%m-\%d \%H:\%M:\%S}.}
   \item{value}{dendrometer value.}
   \item{version}{package version that was used.}
}
\description{
A dataset containing time aligned (\code{L1}) dendrometer data.
}
\keyword{datasets}
