% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dendro_data_L2}
\alias{dendro_data_L2}
\title{Processed Dendrometer Data}
\format{A data frame with 13642 rows and 4 variables.}
\usage{
data(dendro_data_L2)
}
\value{

}
\description{
A dataset containing processed (\code{L2}) dendrometer data.
}
\keyword{datasets}
