% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_treenet.R
\name{download_series}
\alias{download_series}
\title{Download Series from Server}
\usage{
download_series(
  meta_series,
  data_format,
  from,
  to,
  last,
  bind_df,
  reso,
  path_cred,
  server,
  meteo,
  tz,
  use_intl
)
}
\arguments{
\item{meta_series}{\code{data.frame}, contains the names of the series to
download including the reference temperature dataset for each series.}

\item{data_format}{character, select processing level of data. Can either
be \code{"L0"} (i.e. raw data), \code{"L1"} (i.e time-aligned),
\code{"L2"} (i.e. processed), \code{"LM"} (i.e. manually cleaned) or \code{"L2M"} (i.e. L2 and manually
cleaned when available).}

\item{from}{character, optional argument to select data after a specific
date (\code{"YYYY-MM-DD"}).}

\item{to}{character, optional argument to select data up to a specific
date (\code{"YYYY-MM-DD"}).}

\item{last}{numeric, optional argument to specify the number of latest
values. If \code{last} is specified \code{from} and \code{to} need to
be empty.}

\item{bind_df}{logical, indicate whether data should be returned as a
\code{data.frame} (\code{bind_df = TRUE}) or as a named list
\code{bind_df = FALSE}.}

\item{reso}{numeric, desired output time resolution (in minutes). See
Details for more information on data aggregation.}

\item{path_cred}{optional argument to specify the full path to the
file \code{config.yml} containing the database login data. File can
also be copied to the main folder of the package for automatic
recognition.}

\item{server}{character, specify server from which data is downloaded.
Can be either \code{treenet} or \code{decentlab}.}

\item{meteo}{logical, specify whether the site meteo dataset should be
downloaded along with specified data. Automatically enabled or disabled if
\code{"data_format"} ir \code{"L2M"} or \code{"L0"} data, respectively.}

\item{tz}{specify the desired time zone.}

\item{use_intl}{logical, specify whether this function is used
internally on the server. Changes \code{stop()} error messages to
\code{message()} only.}
}
\description{
\code{download_series} downloads specified series from TreeNet or
  Decentlab server
}
\keyword{internal,}
\keyword{treenet}
