% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{executeflagdiff}
\alias{executeflagdiff}
\title{Remove Values With Diff Flag}
\usage{
executeflagdiff(df, length = 2)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{length}{specifies the minimal number of consecutive flags above which
values are overwritten with NA.}
}
\description{
\code{executeflagdiff} removes consecutive values with large differences.
}
\examples{

}
\keyword{internal}
