% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{executeflagout}
\alias{executeflagout}
\title{Remove Outliers}
\usage{
executeflagout(df, len)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{len}{specifies the minimal number of consecutive times the value
difference threshold has to be exceeded for flagging.}
}
\description{
\code{executeflagout} removes outlier values flagged by
  \code{\link{createflagout}}.
}
\examples{

}
\keyword{internal}
