% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{executejump}
\alias{executejump}
\title{Remove Jumps and Outliers}
\usage{
executejump(df)
}
\arguments{
\item{df}{input \code{data.frame}.}
}
\description{
\code{executejump} removes jumps that occur after resetting the dendrometer
  needle and removes outliers.
}
\keyword{internal}
