% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{executejumpout}
\alias{executejumpout}
\title{Remove Jumps and Outliers}
\usage{
executejumpout(df)
}
\arguments{
\item{df}{input \code{data.frame}.}
}
\description{
\code{executejumpout} removes jumps that occur after resetting the dendrometer
  needle and removes outliers.
}
\examples{

}
\keyword{internal}
