% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{fill_na_lead}
\alias{fill_na_lead}
\title{Fills Trailing NA's With First Non-NA Value}
\usage{
fill_na_lead(x)
}
\arguments{
\item{x}{input \code{vector}.}
}
\description{
\code{fill_na_lead} fills leading NA's with first non-NA value.
}
\keyword{internal}
