% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L1.R
\name{fillintergaps}
\alias{fillintergaps}
\title{Interpolate Between Irregular and Regular Timesteps}
\usage{
fillintergaps(df, reso, interpol = reso * 2.1, type = "linear",
  flag = FALSE)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{reso}{desired time resolution of output (in minutes). See
\code{Details} for more information on data aggregation.}

\item{interpol}{numeric, length of gaps (in minutes) in which values are
linearly interpolated.}

\item{type}{specify type of interpolation between regular timesteps.}
}
\description{
\code{fillintergaps} interpolates gaps between irregular and regular
  timesteps, i.e. only over very small gaps (more an alignement than an
  interpolation).
}
\keyword{internal}
