% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L1.R
\name{fillintergaps_prec}
\alias{fillintergaps_prec}
\title{Interpolate Precipitation Data Between Irregular and Regular Timesteps}
\usage{
fillintergaps_prec(df, reso)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{reso}{desired time resolution of output (in minutes). Resolution
should be chosen to be close to the actual measurement intervals.
\code{reso} needs to be a multiple of 5.}
}
\description{
\code{fillintergaps_prec} interpolates gaps between irregular and regular
  timesteps for precipitation data. In contrast to \code{fillintergaps}
  the values are summed up and not lineraly interpolated.
}
\examples{

}
\keyword{internal}
