% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{findmaxmin}
\alias{findmaxmin}
\title{Find Maximum and Minimum in Time Window}
\usage{
findmaxmin(df, reso, st)
}
\arguments{
\item{df}{input \code{data.frame}}

\item{st}{starting row of the first time window.}

\item{span}{width of the time window. Relative to \code{reso}.}

\item{en}{ending row of the last time window.}
}
\description{
\code{findmaxmin} identifies maxima and minima in a specified time window.
  It is a helper function of \code{\link{calcmds}}.
}
\examples{

}
\keyword{internal}
