% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_corr.R
\name{forcejumpnow}
\alias{forcejumpnow}
\title{\code{forcejumpnow} forces a jump in data at specific given date}
\usage{
forcejumpnow(data_L2, force.now)
}
\arguments{
\item{force.now}{character vector: specify the specific date where a jump
will be corrected. There's no period of following days where treenetproc
will look for the largest difference, but the force will be done in the
given date and time specified in \code{force.now}. Dates need to be in a
standard date or datetime format (e.g. \code{\%Y-\%m-\%d \%H:\%M:\%S}).}
}
\description{
\code{forcejumpnow} forces a jump in data at specific given date
}
\keyword{internal}
