% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_data.R
\name{format_input}
\alias{format_input}
\title{Format Input Data}
\usage{
format_input(df, input, tz)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\description{
\code{format_input} formats input data. Wide format is convertet to long
  and correct column classes are assigned.
}
\keyword{internal}
