% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L1.R
\name{generatets}
\alias{generatets}
\title{Generate Regular Time Stamps}
\usage{
generatets(df, reso, all, year, tz)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{reso}{time resolution between two timestamps (in minutes). Resolution
should be chosen to be close to the actual measurement intervals.
\code{reso} needs to be a multiple of 5.}

\item{year}{if \code{year = "full"} then the output \code{data.frame}
contains data of complete years, i.e. \code{\%Y-01-01 to \%Y-12-31}; if
\code{year = "asis"} the output \code{data.frame} contains data over the
same period as the input data.}

\item{tz}{specify the desired time zone. Default is \code{"Etc/GMT-1"}.}
}
\description{
\code{generatets} generates regular time stamps over whole period.
}
\examples{

}
\keyword{internal}
