% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_treenet.R
\name{get_decentlab}
\alias{get_decentlab}
\title{Generate Query for Data from Decentlab Server}
\usage{
get_decentlab(series_single, path_cred = NULL, tz)
}
\arguments{
\item{series_single}{character, name of the sensor to download.}

\item{path_cred}{optional argument to specify the full path to the
file \code{config.yml} containing the database login data. File can
also be copied to the main folder of the package for automatic
recognition.}

\item{tz}{specify the desired time zone.}
}
\description{
\code{get_decentlab} generates a query to download a specified
  series from the Decentlab servers.
}
\keyword{internal,}
\keyword{treenet}
