% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{grostartend}
\alias{grostartend}
\title{Calculate Start and End of Yearly Growth}
\usage{
grostartend(df, tol = 0.05, tz)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{tol}{numeric, defines the amount of yearly growth that needs to be
surpassed for \code{gro_start} to be defined. \code{1 - tol} is the
amount of yearly growth at which \code{gro_end} is defined.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\description{
\code{grostartend} returns the day of the year at which growth starts
  or ends.
}
\keyword{internal}
