% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_seas.R
\name{grow_seas}
\alias{grow_seas}
\title{Calculate the Start and End of the Growing Season}
\usage{
grow_seas(dendro_L2, tol_seas = 0.05, agg_yearly = TRUE, tz = "UTC")
}
\arguments{
\item{dendro_L2}{input \code{data.frame} containing cleaned \code{L2}
dendrometer data (obtained from functions \code{\link{proc_dendro_L2}}
or \code{\link{corr_dendro_L2}}).}

\item{tol_seas}{numeric, defines the amount of yearly growth
that needs to be surpassed before \code{gro_start} is defined.
Likewise, \code{1 - tol_seas} is the amount of yearly growth at which
\code{gro_end} is defined.}

\item{agg_yearly}{logical, specify whether the output is aggregated by year
(\code{agg_yearly = TRUE}) or appended to the \code{L2} data
(\code{agg_yearly = FALSE}).}

\item{tz}{specify the desired time zone.}
}
\value{
The following variables are returned by \code{grow_seas}:
    \item{series_id}{name of the dendrometer series}
    \item{year}{year}
    \item{gro_start}{day of year at which growth starts}
    \item{gro_end}{day of year at which growth ends}

  In case data is not aggregated to yearly values
  (\code{agg_yearly = FALSE}), all columns are appended to the output data.
  The values of \code{gro_start} and \code{gro_end} are only pasted at the
  first timestamp of the year, all other rows are set to \code{NA}.
}
\description{
\code{grow_seas} returns the day of year of growth onset and cessation.
  Values are returned starting from the second year only, since
  the calculation depends on the previous year (see Details for further
  information on the calculation).
}
\details{
\code{gro_start} is defined as the day of year at which the
  maximum dendrometer value of the previous year is surpassed.
  \code{gro_end} is defined as the day of year at which the maximum
  dendrometer value is reached.

  To reduce the influence of potential remaining outliers on
  \code{gro_start} and \code{gro_end}, an adjustable tolerance
  \code{tol_seas} value is used to define growth start and cessation.
  That is, by default \code{gro_start} represents the day of year at which
  5% of yearly growth is surpassed. Likewise, \code{gro_end} represents the
  day of year at which 95% of yearly growth is reached. See Knüsel et al.
  (2020) for a graphical example.
}
\examples{
grow_seas(dendro_L2 = dendro_data_L2)

}
\references{
Knüsel S., Haeni M., Wilhelm M., Peters R.L., Zweifel R. 2020.
  treenetproc: towards a standardized processing of stem radius data.
  In preparation.
}
