% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_variables.R
\name{isdate}
\alias{isdate}
\title{Check Date Format}
\usage{
isdate(datevec, var_name, date_formats, tz)
}
\arguments{
\item{datevec}{character vector in a standard date or datetime format.}

\item{date_formats}{character vector, date or datetime formats that are
checked}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\description{
\code{isdate} checks whether values in a vector are in a standard
  date or datetime format and converts them to \code{POSIXct}.
}
\keyword{internal}
