% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_treenet.R
\name{last_values}
\alias{last_values}
\title{Function to select last values}
\usage{
last_values(df, last, reso, tz)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{last}{numeric, optional argument to specify the number of latest
values. If \code{last} is specified \code{from} and \code{to} need to
be empty.}

\item{reso}{desired time resolution of output (in minutes). See
\code{Details} for more information on data aggregation.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\description{
\code{last_values()} selects specified number of values from current moment
  back in time.
}
\examples{
\dontrun{
last_values(df = treenetdown:::df_last_values, last = 144,
            reso = 10)
}

}
\keyword{internal,}
\keyword{treenet}
