% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_treenet.R
\name{load_credentials}
\alias{load_credentials}
\title{Load Credentials}
\usage{
load_credentials(path_cred = NULL)
}
\arguments{
\item{path_cred}{optional argument to specify the full path to the
file \code{config.yml} containing the database login data. File can
also be copied to the main folder of the package for automatic
recognition.}
}
\description{
\code{load_credentials} loads the credentials from the file 'config.yml'.
}
\keyword{internal,}
\keyword{treenet}
