% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_treenet.R
\name{maxdiff}
\alias{maxdiff}
\title{Calculate Maximum Difference for Time Periods}
\usage{
maxdiff(df, tz)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\description{
\code{maxdiff} calculates the maximum differences in dendrometer
  values over periods of different length.
}
\keyword{internal,}
\keyword{treenet}
