% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\docType{data}
\name{passenv}
\alias{passenv}
\title{Helper Function for Reso}
\format{
An object of class \code{environment} of length 0.
}
\usage{
passenv
}
\arguments{
\item{value}{name of the variable as character string.}
}
\description{
\code{passobj} passes an object to a function call that is saved in a
  separate environment accessible for all functions. Environment is
  saved outside function call to be accessible for all functions.
}
\keyword{internal}
