% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_stats.R
\name{phase_stats}
\alias{phase_stats}
\title{Calculate Parameters of Shrinkage and Expansion Phases}
\usage{
phase_stats(
  dendro_L2,
  phase_wnd = 8,
  plot_phase = FALSE,
  plot_export = TRUE,
  agg_daily = TRUE,
  tz = "UTC"
)
}
\arguments{
\item{dendro_L2}{input \code{data.frame} containing cleaned \code{L2}
dendrometer data (obtained from functions \code{\link{proc_dendro_L2}}
or \code{\link{corr_dendro_L2}}).}

\item{phase_wnd}{numeric, specify the window length (in hours) used to
identify local maxima and minima. A shorter window length leads to the
identification of more local maxima and minima.}

\item{plot_phase}{logical, specify whether identified phases should be
plotted.}

\item{plot_export}{logical, specifies whether the plots are exported as a
\code{PDF} file to the working directory or are plotted in the console.}

\item{agg_daily}{logical, specify whether phase statistics are aggregated
by day (\code{agg_daily = TRUE}) or appended to the \code{L2} data
(\code{agg_daily = FALSE}).}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\value{
The following additional variables are returned by
  \code{phase_stats}:
    \item{day}{day on which the respective phase ends (format
      \code{\%Y-\%m-\%d}.}
    \item{doy}{day of year on which the respective phase ends.}
    \item{shrink_start/exp_start}{timestamp of the start of the shrinkage
      or expansion phase.}
    \item{shrink_end/exp_end}{timestamp of the end of the shrinkage or
      expansion phase.}
    \item{shrink_dur/exp_dur}{duration (minutes) of the shrinkage or
      expansion phase.}
    \item{shrink_amp/exp_amp}{amplitude (\code{µm}) of the shrinkage or
      expansion phase.}
    \item{shrink_slope/exp_slope}{slope (\code{µm * reso^-1})
      of the shrinkage or expansion phase.}
    \item{phase_class}{days are classified as \code{1} if a shrinkage occurs
      during the day, and as \code{-1} if there is an expansion during the
      day. On all other days phase_class is set to NA.}

  In case data is not aggregated to daily values
  (\code{agg_daily = FALSE}), all columns are appended to the input data.
  All parameters related to shrinkage or expansion phases are pasted at
  the timestamp corresponding to the end of the respective phases
  (\code{shrink_end} or \code{exp_end}). All other rows are set to
  \code{NA}.
}
\description{
\code{phase_stats} calculates the timing, duration, amplitude and the rate
  of change of shrinkage and expansion phases. In addition, days on which
  radial change is likely driven by transpiration or temperature are
  identified.
}
\details{
The identification of local maxima and minima in the function
  \code{phase_stats} is inspired by the functionality of
  \code{\link[dendrometeR]{phase_def}} in the package \code{dendrometeR}.
  Overlapping sets of time windows are used to identify local maxima and
  minima, i.e. maxima or minima that appear in both overlapping time windows
  and are not only bound to the start or the end of the respective time
  windows.

  The often reported parameter maximum daily shrinkage (`mds`) can be
  extracted from the output parameters. `mds` is equal to `shrink_amp` in
  case `shrink_dur < 1440` (i.e. the shrinkage is shorter than 24 hours).
  Example code on how to extract `mds` can be found in the vignette
  \href{../doc/Introduction-to-treenetproc.html}{\code{vignette("Introduction-to-treenetproc", package = "treenetproc")}}.
}
\examples{
# Subset dataset for example
library(dplyr)
data_L2 <- dendro_data_L2 \%>\%
   filter(series == "site-1_dendro-2")

phase_stats(dendro_L2 = data_L2, plot_phase = TRUE, plot_export = FALSE)

}
