% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_cycle}
\alias{plot_cycle}
\title{Plot Cycles}
\usage{
plot_cycle(df, cycle, plot_export)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{cycle}{\code{data.frame} containing cycle statistics on the shrinkage
phase, refilling phase and full cycle (as produced by the function
\code{calccycle}.}
}
\value{
Plots are saved to current working directory as
  \code{cycle_plot.pdf}.
}
\description{
\code{plot_cycle} plots maxima and minima selected to calculate cycle
  statistics and prints some characteristics of the cycles.
}
\keyword{internal}
