% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dendro.R
\name{plot_dendro}
\alias{plot_dendro}
\title{Plot Processed Dendrometer Data}
\usage{
plot_dendro(data_L1, data_L2, period = "full", add = TRUE,
  tz = "Etc/GMT-1")
}
\arguments{
\item{data_L1}{time-aligned dendrometer data as produced by \code{proc_L1}.}

\item{data_L2}{processed dendrometer data as produced by
\code{proc_dendro_L2}.}

\item{period}{specify whether plots should be displayed over the whole
period (\code{"full"}) or for each year separately (\code{"yearly"}).}

\item{add}{specify whether \code{L1} data should be plotted along with
\code{L2} data in the second panel.}

\item{tz}{time-zone of provided data. Default is \code{"Etc/GMT-1"}.}
}
\value{
Plots are saved to current working directory as
  \code{processing_L2_plot.pdf}.
}
\description{
\code{plot_dendro()} provides plots of time-aligned (\code{L1}) and
  processed (\code{L2}) dendrometer data to visually assess the processing.
  The first panel shows the \code{L1} data and the second panel the
  processed \code{L2} data. The third panel shows the weekly difference
  between \code{L1} and \code{L2} data. Large differences without an
  apparent jump in the data indicate problems in the processing.
}
\examples{
\dontrun{
plot_dendro(data_L1 = data_L1_dendro, data_L2 = data_L2_dendro,
            period = "yearly", add = TRUE)
}
}
