% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_density}
\alias{plot_density}
\title{Plot density of differences}
\usage{
plot_density(df, low, high, limit_val = 20, frost_thr, reso)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{low}{numeric, low threshold defining outliers. Inherited of the
function \code{\link{calcflagmad}}.}

\item{high}{numeric, high threshold defining outliers. Inherited of the
function \code{\link{calcflagmad}}.}

\item{limit_val}{numeric, defines the x-axis limits of the density plot.
The x-axis limits are drawn at \code{limit_val * threshold}. Threshold
values are inherited of the function \code{\link{calcflagmad}}.}

\item{frost_thr}{numeric, increases the thresholds for outlier
and jump detection in periods of probable frost (i.e. temperature <
\code{lowtemp}). The thresholds are multiplied by the value provided.}

\item{ran}{numeric, range of rows in df considered for the density plot.
Compatible with different window sizes for data processing.}

\item{plot_export}{logical, defines whether plots are exported or shown
directly in the console.}
}
\value{
Plots are saved to current working directory as
  \code{density_plot.pdf}.
}
\description{
\code{plot_density} plots the density of the value differences between
  two time stamps. In addition, the threshold values used to classify
  outliers are shown.
}
\keyword{internal}
