% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_interpol_points}
\alias{plot_interpol_points}
\title{Plot Interpolated Points}
\usage{
plot_interpol_points(df)
}
\arguments{
\item{df}{input \code{data.frame}}
}
\description{
\code{plot_interpol_points} adds points on top of line graph to show
  points that were interpolated.
}
\keyword{internal}
