% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_L1.R
\name{plot_L1}
\alias{plot_L1}
\title{Plot L1 Data}
\usage{
plot_L1(
  dendro_L1,
  dendro_L1_orig = NULL,
  plot_period = "full",
  plot_export = TRUE,
  plot_name = "L1_plot",
  tz = "UTC"
)
}
\arguments{
\item{dendro_L1}{time-aligned dendrometer data as produced by
\code{\link{proc_L1}}.}

\item{dendro_L1_orig}{optional, used in \code{\link{corr_dendro_L1}}. Can be
used to plot a previous version of \code{data_L1} in the background.}

\item{plot_period}{character, specify whether plots should be displayed for
the whole period (\code{plot_period = "full"}), separately for each
year (\code{plot_period = "yearly"}) or for each month
(\code{plot_period = "monthly"}).}

\item{plot_export}{logical, specifies whether the plots are exported as a
\code{PDF} file to the working directory or are plotted in the console.}

\item{plot_name}{character, specify name of the PDF to which the plots are
saved.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\value{
Plots are saved to current working directory as
  \code{L1_plot.pdf} or as specified in \code{plot_name}.
}
\description{
\code{plot_L1} plots time-aligned (\code{L1}) dendrometer or temperature
  data for specified time windows.
}
\examples{
plot_L1(dendro_L1 = dendro_data_L1, plot_period = "monthly",
        plot_export = FALSE)

}
