% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_mds}
\alias{plot_mds}
\title{Plot Maxima and Minima of MDS}
\usage{
plot_mds(df, maxmin, plot_export)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{maxmin}{\code{data.frame} containing maxima and minima used for
the calculation of mds.}
}
\value{
Plots are saved to current working directory as
  \code{mds_plot.pdf}.
}
\description{
\code{plot_mds} plots maxima and minima selected to calculate mds (maximum
  daily shrinkage). Plots are saved as
}
\keyword{internal}
