% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_phase}
\alias{plot_phase}
\title{Plot Phases}
\usage{
plot_phase(df, phase, plot_export)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{phase}{\code{data.frame} containing statistics on the shrinkage
phase and expansion phase.}
}
\value{
Plots are saved to current working directory as
  \code{phase_plot.pdf}.
}
\description{
\code{plot_phase} plots phases of shrinkage and expansion based on
  selected maxima and minima and returns some statistics on the phases.
}
\keyword{internal}
