% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plotting_L1}
\alias{plotting_L1}
\title{Plot Command to Plot L1 Data}
\usage{
plotting_L1(data_L1, data_L1_orig, plot_period, tz)
}
\arguments{
\item{data_L1_orig}{uncorrected original \code{L1} data. If specified, it
is plotted behind the corrected \code{L1} data.}

\item{plot_period}{character, specify whether plots should be displayed for
the whole period (\code{plot_period = "full"}), separately for each
year (\code{plot_period = "yearly"}) or for each month
(\code{plot_period = "monthly"}).}

\item{tz}{specify the desired time zone.}
}
\description{
\code{plotting_L1} plots \code{L1} data.
}
\keyword{internal}
