% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plotting_proc_L2}
\alias{plotting_proc_L2}
\title{Plot Command to Plot L2 Dendrometer Data}
\usage{
plotting_proc_L2(
  data_plot,
  plot_period,
  plot_add = TRUE,
  plot_frost = TRUE,
  plot_interpol = TRUE,
  tz
)
}
\arguments{
\item{data_plot}{input \code{data.frame} containing both \code{L1} and
\code{L2} data as well as changes in \code{L2} for plotting.}

\item{plot_period}{character, specify whether plots should be displayed for
the whole period (\code{plot_period = "full"}), separately for each
year (\code{plot_period = "yearly"}) or for each month
(\code{plot_period = "monthly"}).}

\item{plot_add}{logical, specify whether \code{L1} data should be plotted
along with \code{L2} data in the second panel of the plot.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\description{
\code{plotting_proc_L2} contains the code necessary to plot \code{L2}
  dendrometer data.
}
\keyword{internal}
