% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_server.R
\name{proc_server}
\alias{proc_server}
\title{Process Treenet Data to L1 or L2 on Server}
\usage{
proc_server(
  from = NULL,
  to = NULL,
  last = NULL,
  data_version = NULL,
  frost_thr = 5,
  lowtemp = 5,
  tol_jump = 50,
  tol_out = 10,
  interpol = NULL,
  frag_len = NULL,
  proc_to = "L2",
  path_cred = NULL,
  year = "asis",
  iter_clean = 1,
  tz = "Etc/GMT-1"
)
}
\arguments{
\item{from}{character, optional argument to select data after a specific
date (\code{"YYYY-MM-DD"}).}

\item{to}{character, optional argument to select data up to a specific
date (\code{"YYYY-MM-DD"}).}

\item{last}{numeric, optional argument to specify the number of latest
values. If \code{last} is specified \code{from} and \code{to} need to
be empty.}

\item{data_version}{character, optional argument specifying the package version
number. If provided, \code{L2} data generated by the package version
specified is downloaded from the server (download of \code{L0} or
\code{L1} data is not affected).}

\item{frost_thr}{numeric, increases the thresholds for outlier
classifiation in periods of probable frost (i.e. temperature <
\code{lowtemp}). The thresholds are multiplied by the value provided.}

\item{lowtemp}{numeric, specifies temperature in °C below which shrinkage
in stem diameter due to frost is expected. Default value is set to
\code{5°C} due to hysteresis shortly before or after frost events.}

\item{tol_jump}{numeric, defines the tolerance of the threshold above or
below which a value is flagged for jump correction.}

\item{tol_out}{numeric, defines the tolerance of the threshold above or
below which a value is classified as outlier (see Details for further
information).}

\item{interpol}{numeric, length of gaps (in minutes) in which values are
linearly interpolated. Set \code{interpol = 0} to disable gapfilling.
If \code{interpol = NULL} the default value is set to
\code{interpol = 2.1 * reso}.}

\item{frag_len}{numeric, specifies the length of data fragments occurring
in-between missing data that are automatically deleted during data
cleaning. This can be helpful to remove short fragments of erroneous data
within a period of missing data, i.e. after jumps. If
\code{frag_len = NULL} the devault value is set to \code{frag_len = 2.1}.}

\item{path_cred}{optional argument to specify the full path to the
file \code{config.yml} containing the database login data. File can
also be copied to the main folder of the package for automatic
recognition.}

\item{year}{if \code{year = "full"} then the output \code{data.frame}
contains data of complete years, i.e. \code{YYYY-01-01 to YYYY-12-31}; if
\code{year = "asis"} the output \code{data.frame} covers the same period
as the input data.}

\item{iter_clean}{numeric, specifies the number of times the cleaning
process is repeated. Can be used to check whether running the cleaning
process many times has an effect on the results.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\value{
The function returns a \code{data.frame} with processed dendrometer
 data containing the following columns:
   \item{series}{name of the series.}
   \item{ts}{timestamp with format \code{\%Y-\%m-\%d \%H:\%M:\%S}.}
   \item{value}{dendrometer value.}
   \item{max}{highest measured value up to this timestamp.}
   \item{twd}{tree water deficit, i.e. the amount of stem shrinkage
     expressed as the difference between \code{max} and \code{value}.}
   \item{mds}{maximum daily shrinkage, calculated as the difference between
     a local maximum that occurs before a local minimum during one day. If
     there is no local maximum or minimum or if the minimum occurs prior to
     the maximum, then \code{mds = NA}. This may occur on days with rain or
     in winter (see Details for further information).}
   \item{gro_yr}{growth since the beginning of the year. Also calculated
     for years with missing data.}
   \item{gro_start}{day of year at which growth starts. \code{gro_start} is
     defined as the day of year at which 5\% of total yearly growth is
     surpassed.}
   \item{gro_end}{day of year at which growth stops. \code{gro_end} is
     defined as the day of year at which 95\% of total yearly growth is
     reached.}
   \item{flags}{character vector specifying the changes that occurred
     during the processing. For more details see the following vignette:
     \href{../doc/Introduction-to-treenetproc.html}{\code{vignette("Introduction-to-treenetproc", package = "treenetproc")}}}
   \item{version}{package version that was used.}
}
\description{
\code{proc_server} processes \code{L0} dendrometer data on the
  treenet server to \code{L1} or \code{L1} data to \code{L2}.
  For the processing to \code{L1}, all available series are processed.
  For the processing to \code{L2}, all dendrometer series are processed
  with the respective reference temperature datasets.
}
\keyword{server}
\keyword{treenet,}
