% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_treenet_L2.R
\name{proc_treenet_L2}
\alias{proc_treenet_L2}
\title{Process Treenet Data to L2}
\usage{
proc_treenet_L2(site = NULL, sensor_name = NULL, from = NULL,
  to = NULL, path_cred = NULL, temp_name = NULL, reso = 10,
  year = "asis", val_range = c(0, 20000), diffwin = 2000,
  diffsum = 1000, lowtemp = 5, tz = "Etc/GMT-1", plot = TRUE,
  period = "full", add = TRUE)
}
\arguments{
\item{site}{character string specifying the site. String is not
case sensitive.}

\item{sensor_name}{character string specifying the name of a single or
multiple sensors (e.g. \code{"LWF-Demo-1.dendrometer.ch0"}) of the
same site.}

\item{from}{optional argument to select data after a specific date
(\code{"YYYY-MM-DD"}).}

\item{to}{Optional argument to select data up to a specific date
(\code{"YYYY-MM-DD"}).}

\item{path_cred}{optional argument to specify the full path to the
file \code{config.yml} containing the database login data. File can
also be copied to the main folder of the package for automatic
recognition.}

\item{temp_name}{optional argument to specify the series name of a
temperature dataset. Needed if there are no or multiple on-site air
temperature measurements. If \code{temp_name} is not specified and no
temperature dataset is found on the server, a sample dataset will be
used with a warning.}

\item{reso}{desired time resolution of output (in minutes). Resolution
should be chosen to be close to the actual measurement intervals.
\code{reso} needs to be a multiple of 5.}

\item{year}{if \code{year = "full"} then the output \code{data.frame}
contains data of complete years, i.e. \code{YYYY-01-01 to YYYY-12-31}; if
\code{year = "asis"} the output \code{data.frame} covers the same period
as the input data.}

\item{val_range}{numeric vector specifying the minimum and maximum
\code{c(min, max)} of credible dendrometer measurement values. Values
lower than \code{min} or higher than \code{max} are deleted without
notice.}

\item{diffwin}{maximal hourly difference expected in winter.}

\item{diffsum}{maximal hourly difference expected in summer.}

\item{lowtemp}{numeric, specifies temperature in °C below which shrinkage
in stem diameter due to frost is expected.}

\item{tz}{specify the desired time zone. Default is \code{"Etc/GMT-1"}.}

\item{plot}{logical, specify whether a comparison of \code{L1} and \code{L2}
data should be plotted.}

\item{period}{specify whether plots should be displayed over the whole
period (\code{period = "full"}) or for each year separately
(\code{period = "yearly"}).}

\item{add}{logical, specify whether \code{L1} data should be plotted along
with \code{L2} data in the second panel of the plot.}
}
\value{
The function returns:
 a \code{data.frame} with processed dendrometer data containing the
 following columns
   \item{series}{name of the series.}
   \item{ts}{timestamp with format \code{\%Y-\%m-\%d \%H:\%M:\%S}.}
   \item{value}{dendrometer value.}
   \item{max}{highest measured value up to this timestamp.}
   \item{twd}{tree water deficit, i.e. the amount of stem shrinkage
     expressed as the difference between \code{max} and \code{value}.}
   \item{mds}{maximum daily shrinkage, calculated as the difference between
     a local maximum that occurs before a local minimum during one day. If
     there is no local maximum or minimum or if the minimum occurs prior to
     the maximum, then \code{mds = NA}. This may occur on days with rain or
     in winter.}
   \item{gro_yr}{growth since the beginning of the year. Also calculated if
     for years with missing data.}
   \item{flags}{number specifying whether and which changes occurred during
     the processing.}
   \item{version}{processing version.}
}
\description{
\code{proc_treenet_L2()} processes \code{L0} dendrometer data from the
  treenet server directly to \code{L2}.
}
\examples{
\dontrun{
proc_treenet_L2(site = "sagno", year = "full", period = "yearly")

proc_treenet_L2(site = "lens", add = FALSE)

proc_treenet_L2(sensor_name = c("Alvaneu-2.dendrometer.ch0",
                "Alvaneu-4.dendrometer.ch0"))

# throws error
proc_treenet_L2(sensor_name = c("Alvaneu-2.dendrometer.ch0",
                "Alvaneu-4.dendrometer.ch0", "Bachtel-2.dendrometer.ch0"))
}
}
\seealso{
\code{\link{proc_L1}} to process dendrometer or climate data to
  \code{L1} only (i.e. time-aligned data), \code{\link{proc_dendro_L2}}
  to process dendrometer data from \code{L1} to \code{L2}.
}
