% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalc_growth_variables.R
\name{recalc_growth_variables}
\alias{recalc_growth_variables}
\title{Re-calculate the growth variables: max, twd and gro_yr}
\usage{
recalc_growth_variables(
  dendro_L2,
  tz = "UTC",
  plot = FALSE,
  plot_export = FALSE
)
}
\arguments{
\item{dendro_L2}{input \code{data.frame} containing cleaned \code{L2}
dendrometer data (obtained from functions \code{\link{proc_dendro_L2}}
or \code{\link{corr_dendro_L2}}).}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}

\item{plot}{logical, specify whether the changes that occurred during data
cleaning should be plotted.}

\item{plot_export}{logical, specifies whether the plots are exported as a
\code{PDF} file to the working directory or are plotted in the console.}
}
\value{
The function returns a \code{data.frame} with the \code{treenetproc}
growth output (max, twd and gro_yr) recalculated. It leaves the other columns
unmodified. This function is useful if you has changed the initial moment of
measure (for instance, by filtering the data) and you want to recalculate the
growing stats or phase stats, among others.
}
\description{
\code{recalc_growth_variables}
}
\examples{
recalc_growth_variables(dendro_L2 = dendro_data_L2, tz = 'Europe/Madrid')

}
\seealso{
\code{\link{grow_seas}} to calculate the growing season variables
   from the dendro_L2 data, which is updated with this function.
}
