% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_input_data.R
\name{remove_lead_trail_na}
\alias{remove_lead_trail_na}
\title{Remove Leading and Trailing NA's}
\usage{
remove_lead_trail_na(df)
}
\arguments{
\item{df}{input \code{data.frame}}
}
\value{
list, first element is \code{df} and the second element are the
  leading and trailing \code{NA}.
}
\description{
\code{lead_trail_na} removes and saves leading and trailing \code{NA} to a
  \code{list}.
}
\keyword{internal}
