% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{removeconsec}
\alias{removeconsec}
\title{Remove Consecutive Values}
\usage{
removeconsec(df, remove, notremove)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{remove}{numeric, in which consecutive values are removed
(i.e. maxima or minima).}

\item{notremove}{numeric, in which consecutive values are not removed
(i.e. maxima or minima).}
}
\description{
\code{removeconsec} removes consecutive maxima or minima and keeps only
  the higher or lower value, respectively. The function makes sure no two
  maxima or minima appear consecutively. It is a helper function of
  \code{\link{calcmds}}.
}
\examples{

}
\keyword{internal}
