% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_corr.R
\name{removecorr}
\alias{removecorr}
\title{Removes Wrong Corrections in Data}
\usage{
removecorr(data_L1, data_L2, remove, tz)
}
\arguments{
\item{remove}{numeric vector, specify numbers of the changes that should
be removed. Numbers are displayed in the plots produced by
\code{\link{proc_dendro_L2}} or \code{\link{plot_proc_L2}} (numbers
are only displayed if \code{plot_period = "monthly"}).}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\description{
\code{removecorr} removes wrong corrections in the dendrometer data.
}
\keyword{internal}
