% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L3.R
\name{removejump}
\alias{removejump}
\title{Removes Wrong Jump in Data}
\usage{
removejump(data_L1, data_L2, remove)
}
\arguments{
\item{data_L1}{time-aligned dendrometer data as produced by
\code{\link{proc_L1}}. Optional, only needed for \code{remove} and if
\code{plot = TRUE}.}

\item{data_L2}{processed dendrometer data as produced by
\code{\link{proc_dendro_L2}}.}

\item{remove}{character vector, specify the dates after which jumps
should be removed. The first jump that occurs after the date
specified is removed. If multiple jumps occur within \code{10} timesteps,
all of these jumps are removed. Dates can be formatted as
\code{"YYYY-MM-DD"} or \code{"YYYY-MM-DD HH:MM:SS"}.}
}
\description{
\code{removejump} removes wrong corrections in the dendrometer data.
}
\examples{

}
\keyword{internal}
