% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{removeoutliers}
\alias{removeoutliers}
\title{Removes Outliers Based on Quantile}
\usage{
removeoutliers(df, quan = 0.001, wnd = 3, reso, span = 60/reso * 24 *
  (wnd/2), by = 60/reso * 24)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{quan}{quantile that identifies outliers.}

\item{span}{size of window for which quantiles are calculated.}

\item{by}{spacing between time windows for quantile calculation.}
}
\description{
\code{removeoutliers} removes outliers that are above or below a specified
  quantile within a defined window. (used to be implemented in createflags1)
}
\examples{

}
\keyword{internal}
