% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_input_data.R
\name{reso_check}
\alias{reso_check}
\title{Check Time Resolution of Input Data}
\usage{
reso_check(df)
}
\arguments{
\item{df}{input \code{data.frame}.}
}
\description{
\code{reso_check} checks time resolution of input \code{data.frame}.
}
\examples{

}
\keyword{internal}
