% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_data.R
\name{reso_check_L1}
\alias{reso_check_L1}
\title{Check Time Resolution of L1 Input Data}
\usage{
reso_check_L1(df)
}
\arguments{
\item{df}{input \code{data.frame}.}
}
\description{
\code{reso_check} checks time resolution of \code{L1} input
  \code{data.frame}.
}
\keyword{internal}
