% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_corr.R
\name{reversecorr}
\alias{reversecorr}
\title{Reverses Wrong Corrections in Data}
\usage{
reversecorr(data_L1, data_L2, reverse, tz)
}
\arguments{
\item{reverse}{numeric vector, specify numbers of the changes that should
be reversed. Numbers are reported in the plots produced by
\code{\link{proc_dendro_L2}} or \code{\link{plot_proc_L2}} with the
argument \code{plot_period = "monthly"}.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\description{
\code{reversecorr} reverses wrong corrections in the dendrometer data.
}
\keyword{internal}
