% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_corr.R
\name{revflags}
\alias{revflags}
\title{Reverse flags}
\usage{
revflags(df)
}
\arguments{
\item{df}{input \code{data.frame}.}
}
\description{
\code{revflags} removes all the flag of changes done
by any treenetproc automatic correction and adds instead
the "rev" flag
}
\keyword{internal}
