% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L1.R
\name{roundtimetoreso}
\alias{roundtimetoreso}
\title{Rounds Irregular Time Stamp to Regular}
\usage{
roundtimetoreso(df, reso, pos, tz)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{reso}{desired time resolution of output (in minutes). See
\code{Details} for more information on data aggregation.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\description{
\code{roundtimetoreso} rounds first and last irregular timestep to
  next or previous regular timestep.
}
\keyword{internal}
