% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{saveplotthr}
\alias{saveplotthr}
\title{Save Applied Cleaning Thresholds for Plotting}
\usage{
saveplotthr(df, thr_out, thr_jump)
}
\arguments{
\item{thr_out}{vector, containing the values of the minimum and maximum
outlier thresholds applied.}

\item{thr_jump}{vector, containing the values of the minimum and maximum
jump thresholds applied.}
}
\description{
\code{saveplotthr} stores the applied threshold values for jump and
  outlier detection during data cleaning. The values are later used
  for plotting (i.e. presented below the plot with the yearly growth
  curves).
}
\keyword{internal}
