% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_data.R
\name{select_data}
\alias{select_data}
\title{Select Dendrometer and Temperature Data from Treenet Server}
\usage{
select_data(site = NULL, sensor_class = NULL, sensor_name = NULL,
  data_format = NULL, from = NULL, to = NULL, path_cred = NULL,
  temp_name = NULL, select_temp = TRUE, last = NULL,
  export = FALSE, bind_df = FALSE, server = "treenet", reso = 10,
  tz = "Etc/GMT-1")
}
\arguments{
\item{site}{character, specify the site. String is not case sensitive.}

\item{sensor_class}{charcter, specify a single or multiple sensor
classes. String does not have to contain the full name of the sensor
class, i.e. \code{"dendro"} works for \code{"dendrometer"}. String is not
case sensitive.}

\item{sensor_name}{character, specify the name of a single or
multiple sensors (e.g. \code{"LWF-Demo-1.dendrometer.ch0"}) of the
same site.}

\item{data_format}{character, select processing level of data. Can either
be \code{"L0"} (i.e. raw data), \code{"L1"} (i.e time-aligned) or
\code{"L2"} (i.e. processed).}

\item{from}{character, optional argument to select data after a specific
date (\code{"YYYY-MM-DD"}).}

\item{to}{character, optional argument to select data up to a specific
date (\code{"YYYY-MM-DD"}).}

\item{path_cred}{optional argument to specify the full path to the
file \code{config.yml} containing the database login data. File can
also be copied to the main folder of the package for automatic
recognition.}

\item{temp_name}{character, optional argument to specify the series name
of a temperature dataset. Needed if there are no or multiple on-site air
temperature measurements. If \code{temp_name} is not specified and no
temperature dataset is found on the server, a sample dataset will be
used with a warning.}

\item{select_temp}{logical, specifies whether a temperature dataset should
be automatically selected if the respective site contains temperature
measurements.}

\item{last}{numeric, optional argument to specify the number of latest
values. If \code{last} is specified \code{from} and \code{to} need to
be empty.}

\item{export}{logical, indicate whether each sensor will be saved as
an \code{.RData}-file in the current working directory
(\code{export = TRUE}) or will be returned to the console.}

\item{bind_df}{logical, indicate whether data should be returned as a
\code{data.frame} (\code{bind_df = TRUE}) or as a named list
\code{bind_df = FALSE}.}

\item{server}{character, specify server from which data is downloaded.
Can be either \code{treenet} or \code{decentlab}.}

\item{reso}{desired time resolution of output (in minutes). See
\code{Details} for more information on data aggregation.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\value{
The output will be one \code{.RData} file for each sensor in case
\code{export = TRUE}. Otherwise data will be saved in a list.
}
\description{
\code{select_data} accesses and downloads dendrometer and temperature
  data from treenet server. Function imported and adapted from package
  \code{treenetdown}.
}
\examples{
\dontrun{
select_data(site = "birmensdo")
}

}
\keyword{internal,}
\keyword{treenet}
