% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_treenet.R
\name{select_ref_data}
\alias{select_ref_data}
\title{Select Reference Data for Selected Series}
\usage{
select_ref_data(meta_list)
}
\arguments{
\item{meta_list}{list, containing the metadata file (first element) and
all series specified for download from the server (second element).}
}
\description{
\code{select_ref_data} selects reference temperature data as well as outlier, jump, and frost tolerances for dendrometer
  series based on the metadata file.
}
\keyword{internal,}
\keyword{treenet}
