% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_treenet.R
\name{select_series}
\alias{select_series}
\title{Select Series from Metadata File}
\usage{
select_series(site, sensor_class, sensor_name, path_cred)
}
\arguments{
\item{site}{character, specify the site. String is not case sensitive.}

\item{sensor_class}{charcter, specify a single or multiple sensor
classes. String does not have to contain the full name of the sensor
class, i.e. \code{"dendro"} works for \code{"dendrometer"}. String is not
case sensitive.}

\item{sensor_name}{character, specify the name of a single or
multiple sensors (e.g. \code{"LWF-Demo-1.dendrometer.ch0"}).}

\item{path_cred}{optional argument to specify the full path to the
file \code{config.yml} containing the database login data. File can
also be copied to the main folder of the package for automatic
recognition.}
}
\description{
\code{select_series} selects series from the metadata file based on
  selection criteria specified by the user (i.e. site, sensor_class, etc.).
  In addition, reference temperature data is selected for all dendrometer
  series.
}
\keyword{internal,}
\keyword{treenet}
