% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_treenet.R
\name{select_temp_data}
\alias{select_temp_data}
\title{Select Reference Temperature Data for Selected Series}
\usage{
select_temp_data(meta_list)
}
\arguments{
\item{meta_list}{list, containing the metadata file (first element) and
all series specified for download from the server (second element).}
}
\description{
\code{select_temp_data} selects reference temperature data for dendrometer
  series based on the metadata file.
}
\keyword{internal,}
\keyword{treenet}
