% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{summariseflags}
\alias{summariseflags}
\title{Summarise Flags}
\usage{
summariseflags(df)
}
\arguments{
\item{df}{input \code{data.frame}.}
}
\description{
\code{summariseflags} summarises all previously created flags in one column.
}
\keyword{internal}
