% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temp_data_L0}
\alias{temp_data_L0}
\title{Raw Temperature data}
\format{
A data frame with 145770 rows and 3 variables.
\describe{
  \item{series_id}{name of the data series}
  \item{ts}{timestamp of the measurements}
  \item{value}{measured values}
}
}
\usage{
data(temp_data_L0)
}
\description{
A dataset containing raw temperature data that is formatted in long
format.
}
\keyword{datasets}
