% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treenetproc.R
\docType{package}
\name{treenetproc}
\alias{treenetproc}
\title{treenetproc - Clean, process and visualize dendrometer data}
\description{
The \code{treenetproc} package contains three main functions:
  \code{proc_L1}, \code{proc_dendro_L2}, \code{corr_dendro_L2}
}
\section{\code{proc_L1()}}{

  The function \code{proc_L1} aligns raw dendrometer and temperature
  measurements to regular time intervals.
}

\section{\code{proc_dendro_L2()}}{

  The function \code{proc_dendro_L2} removes outliers and jumps or shifts
  in time-aligned dendrometer data.
}

\section{\code{corr_dendro_L2()}}{

  The function \code{corr_dendro_L2} can be used to manually correct
  remaining errors after data cleaning with \code{proc_dendro_L2}.
}

\section{Vignettes}{

  More information and some examples on how to use the functions can be
  found in the package vignette. To see the vignette use:\cr
  \code{browseVignettes("treenetproc")}\cr
  If \code{browseVignettes} does not work, the package may need to be
  reinstalled along with the vignette:\cr
  \code{library(devtools)}\cr
  \code{devtools::install_github("treenet/treenetproc",
  build_vignettes = TRUE))}
}

