% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_input_data.R
\name{ts_overlap_check}
\alias{ts_overlap_check}
\title{Check for Time Overlap Between Datasets}
\usage{
ts_overlap_check(df, tem)
}
\arguments{
\item{df}{input \code{data.frame} of dendrometer data.}

\item{tem}{input \code{data.frame} of temperature or climate data.}
}
\description{
\code{ts_overlap_check} checks whether measurement periods of dendrometer
  data and temperature data overlap.
}
\examples{

}
\keyword{internal}
