% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L1.R
\name{tsalign_prec}
\alias{tsalign_prec}
\title{Time-Alignement of Precipitation Data to Regular Time Intervals}
\usage{
tsalign_prec(df, reso, year, tz)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{reso}{desired time resolution of output (in minutes). Resolution
should be chosen to be close to the actual measurement intervals.
\code{reso} needs to be a multiple of 5.}

\item{year}{if \code{year = "full"} then the output \code{data.frame}
contains data of complete years, i.e. \code{YYYY-01-01 to YYYY-12-31}; if
\code{year = "asis"} the output \code{data.frame} covers the same period
as the input data.}

\item{tz}{specify the desired time zone. Default is \code{"Etc/GMT-1"}.}
}
\description{
\code{tsalign_prec} aligns precipitation measurements to regular time
  intervals by summing up values of irregular time intervals.
}
\examples{

}
\keyword{internal}
