% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{axis_labels_period}
\alias{axis_labels_period}
\title{Define Axis Labels Based on Period}
\usage{
axis_labels_period(df, plot_period, tz)
}
\arguments{
\item{df}{input \code{data.frame}}

\item{plot_period}{specify whether plots should be displayed over the whole
plot_period (\code{plot_period = "full"}), for each year separately
(\code{plot_period = "yearly"}) or for each month
(\code{plot_period = "monthly"}).}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\description{
\code{axis_labels_period} defines axis ticks and labels based on the
  period used for plotting.
}
\keyword{internal}
