% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{calcflagmad}
\alias{calcflagmad}
\title{Define Outliers Based on Value Distribution}
\usage{
calcflagmad(df, reso, wnd = 6, n_mad = 9, frost,
  plot_density = FALSE)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{frost}{logical, defines whether outliers should be calculated for
frost periods or other periods.}

\item{plot_density}{logical, defines whether density plots should be drawn
in the console. Can be used to check outlier thresholds.}
}
\description{
\code{calcflagmad} is a helper function of \code{\link{createflagmad}}.
}
\examples{

}
\keyword{internal}
