% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{calcflagmad}
\alias{calcflagmad}
\title{Define Outliers Based on Value Distribution}
\usage{
calcflagmad(df, reso, wnd = NULL, tol = 10, frost, frost_thr,
  save_thr = FALSE, correction = NULL)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{frost}{logical, defines whether outliers should be calculated for
frost periods or other periods.}

\item{frost_thr}{numeric, increases the thresholds for outlier
classifiation in periods of probable frost (i.e. temperature <
\code{lowtemp}). The thresholds are multiplied by the value provided.}

\item{save_thr}{logical, specifies whether the applied thresholds are
saved for later plotting or not.}
}
\description{
\code{calcflagmad} is a helper function of \code{\link{createflagmad}}.
}
\keyword{internal}
