% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{calcmds}
\alias{calcmds}
\title{Calculate Maximum Daily Shrinkage (MDS)}
\usage{
calcmds(df, tz)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{tz}{specify the desired time zone. Default is \code{"Etc/GMT-1"}.}
}
\description{
\code{calcmds} calculates the maximum daily shrinkage (mds). Mds is defined
  as the largest daily shrinkage. First, local maxima and minima are
  identified using a moving window. Mds is only calculated if a local
  maximum occurs before a local minimum (i.e. if the stem shrinks during
  the day).
}
\examples{

}
\keyword{internal}
