% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{calcmds}
\alias{calcmds}
\title{Calculate Maximum Daily Shrinkage (MDS)}
\usage{
calcmds(df, tz, reso, plot_mds = FALSE)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{tz}{specify the desired time zone. Default is \code{"Etc/GMT-1"}.}

\item{plot_mds}{logical, specify whether maxima and minima used for the
calculation of mds (maximum daily shrinkage) should be plotted.}
}
\description{
\code{calcmds} calculates the maximum daily shrinkage (mds). Mds is defined
  as the largest daily shrinkage. First, local maxima and minima are
  identified using a moving window. Mds is only calculated if a local
  maximum occurs before a local minimum (i.e. if the stem shrinks during
  the day).
}
\details{
\code{calcmds} is inspired by the function
  \code{\link[dendrometeR]{phase_def}} in the package \code{dendrometeR}.
}
\examples{

}
\keyword{internal}
