% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{calcmissing}
\alias{calcmissing}
\title{Calculates Percentages of Interpolated, Deleted and Missing Data}
\usage{
calcmissing(data_L1, data_L2)
}
\arguments{
\item{data_L1}{time-aligned dendrometer data as produced by
\code{\link{proc_L1}}.}

\item{data_L2}{processed dendrometer data as produced by
\code{\link{proc_dendro_L2}}.}
}
\value{
a list of length = 3 with percentages of interpolated, deleted,
  and missing data.
}
\description{
\code{calcmissing} calculates the percentage of interpolated, deleted and
  missing data.
}
\keyword{internal}
