% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_variables.R
\name{check_datevec}
\alias{check_datevec}
\title{Check Date Vector Input}
\usage{
check_datevec(datevec, tz)
}
\arguments{
\item{datevec}{character vector in a standard date or datetime format.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\description{
\code{check_datevec} checks whether all dates specified are in a standard
  date or datetime format. Dates are converted to \code{POSIXct}.
}
\keyword{internal}
