% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_variables.R
\name{check_series}
\alias{check_series}
\title{Checks Input of Series}
\usage{
check_series(df, series)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{series}{character, specify the name of a single dendrometer series
for which changes should be made. Data of other series is left unchanged.
Not needed if only a single series is provided.}
}
\description{
\code{check_series} checks the input of the variabel series.
}
\keyword{internal}
