% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_dendro_L1.R
\name{corr_dendro_L1}
\alias{corr_dendro_L1}
\title{Manually Correct Time-Aligned Dendrometer Data}
\usage{
corr_dendro_L1(data_L1, delete, series = NULL, plot = TRUE,
  tz = "Etc/GMT-1")
}
\arguments{
\item{data_L1}{time-aligned dendrometer data as produced by
\code{\link{proc_L1}}.}

\item{delete}{character vector, specify pairs of dates between which
all dendrometer data will be deleted (i.e. 4 dates will result in two
periods: 1-2 and 3-4 in which data is deleted). Dates need to be in a
standard date or datetime format (e.g. \code{\%Y-\%m-\%d \%H:\%M:\%S}).}

\item{series}{character, specify the name of a single dendrometer series
for which changes should be made. Data of other series is left unchanged.
Not needed if only a single series is provided.}

\item{plot}{logical, plots the corrected dataset with the original data
in the back.}

\item{tz}{specify the desired time zone. Default is \code{"Etc/GMT-1"}.}
}
\value{

}
\description{
\code{corr_dendro_L1} deletes values in periods with erroneous data of
  time-aligned (\code{L1}) dendrometer or climate data. Values are
  overwritten with \code{NA} without further notice.
}
\examples{

}
