% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_dendro_L3.R
\name{corr_dendro_L3}
\alias{corr_dendro_L3}
\title{Manually Correct Processed Dendrometer Data}
\usage{
corr_dendro_L3(data_L1 = NULL, data_L2, remove = NULL, force = NULL,
  delete = NULL, series = NULL, n_days = 5, plot = TRUE,
  tz = "Etc/GMT-1")
}
\arguments{
\item{data_L1}{time-aligned dendrometer data as produced by
\code{\link{proc_L1}}. Optional, only needed for \code{remove} and if
\code{plot = TRUE}.}

\item{data_L2}{processed dendrometer data as produced by
\code{\link{proc_dendro_L2}}.}

\item{remove}{numeric vector, specify numbers of the changes that should
be removed. Numbers are displayed in the plots produced by
\code{\link{proc_dendro_L2}} or \code{\link{plot_proc_L2}} (numbers are
only displayed if \code{plot_period = "monthly"}).}

\item{force}{character vector, specify the dates after which jumps
should occur. The jump with the largest value difference occurring in a
plot_period of \code{n_days} after the specified dates in \code{force} is
corrected. Dates need to be in a standard date or
datetime format (e.g. \code{\%Y-\%m-\%d \%H:\%M:\%S}).}

\item{delete}{character vector, specify pairs of dates between which
all dendrometer data will be deleted (i.e. 4 dates will result in two
periods: 1-2 and 3-4 in which data is deleted). Dates need to be in a
standard date or datetime format (e.g. \code{\%Y-\%m-\%d \%H:\%M:\%S}).}

\item{series}{character, specify the name of a single dendrometer series
for which changes should be made. Data of other series is left unchanged.
Not needed if only a single series is provided.}

\item{n_days}{numeric, length of the period (in days) following the dates
specified in \code{force} in which a jump is looked for.}

\item{plot}{logical, specify whether changes between \code{L2} and
\code{L3} should be plotted.}

\item{tz}{specify the desired time zone. Default is \code{"Etc/GMT-1"}.}
}
\value{
The function returns a \code{data.frame} with corrected \code{L2}
  dendrometer data.
}
\description{
\code{corr_dendro_L3} corrects errors in processed \code{L2} dendrometer
  data that are easier to correct by hand than by changing parameter values
  in \code{\link{proc_dendro_L2}}.
}
\examples{

}
