% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{createflagfragment}
\alias{createflagfragment}
\title{Create Flag for Small Data Fragments}
\usage{
createflagfragment(out, frag_len = interpol/passobj("reso"))
}
\arguments{
\item{out}{logical, vector with outlier flags.}

\item{frag_len}{numeric, specifying the length of the fragments that are
flagged.}
}
\description{
\code{createflagfragment} flags short fragments of not flagged data that
  is in-between flagged data. Such cases occur often in spikes of outlier
  data.
}
\keyword{internal}
