% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{createflagmad}
\alias{createflagmad}
\title{Creates Flag for Outliers Based on Median Absolute Deviation}
\usage{
createflagmad(df, reso, wnd, n_mad, plot_density)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{n_mad}{numeric, defines the thresholds above or below which values
are classified as outliers (see Details for further information).}

\item{plot_density}{logical, defines whether density plots should be drawn
in the console. Can be used to check outlier thresholds.}
}
\description{
\code{createflagmad} creates flag for outliers that are above or below a
  threshold distant from the first or third quantile. The threshold is
  specified by the the median absolute deviation
  (\code{\link[stats]{mad}}).
}
\examples{

}
\keyword{internal}
