% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{createflagmad}
\alias{createflagmad}
\title{Creates Flag for Outliers Based on Median Absolute Deviation}
\usage{
createflagmad(df, reso, wnd, tol, save_thr, frost_thr, correction)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{save_thr}{logical, specifies whether the applied thresholds are
saved for later plotting or not.}

\item{frost_thr}{numeric, increases the thresholds for outlier
classifiation in periods of probable frost (i.e. temperature <
\code{lowtemp}). The thresholds are multiplied by the value provided.}
}
\description{
\code{createflagmad} creates flag for outliers that are above or below a
  threshold distant from the first or third quantile. The threshold is
  specified by the the median absolute deviation
  (\code{\link[stats]{mad}}).
}
\keyword{internal}
