% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{createflagout}
\alias{createflagout}
\title{Create Flag for Outliers}
\usage{
createflagout(df, out, len)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{len}{specifies the minimal number of consecutive times the value
difference threshold has to be exceeded for flagging.}
}
\description{
\code{createflagout} creates flags for outlier values.
}
\keyword{internal}
