% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L2.R
\name{createjumpflag}
\alias{createjumpflag}
\title{Creates Flag for Jumps and Outliers in Data}
\usage{
createjumpflag(df, thr = 0.2)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{thr}{specifies the threshold to discriminate between outliers and
jumps due to adjustments of the dendrometer needle.}
}
\description{
\code{createjumpflag} creates flag for jumps and ouliers. Outliers can be
  single points or errors before, after or in between jumps.
}
\keyword{internal}
