% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_corr.R
\name{deleteperiod}
\alias{deleteperiod}
\title{Deletes Data in Specified Period}
\usage{
deleteperiod(df, delete)
}
\arguments{
\item{df}{input \code{data.frame}.}
}
\description{
\code{deleteperiod} deletes dendrometer data in specified period.
}
\keyword{internal}
