% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dendro_data_L2}
\alias{dendro_data_L2}
\title{Cleaned Dendrometer Data}
\format{A data frame with 151236 rows and 4 variables.}
\usage{
data(dendro_data_L2)
}
\value{
The function returns a \code{data.frame} with processed dendrometer
  data containing the following columns:
   \item{series}{name of the dendrometer series.}
   \item{ts}{timestamp with format \code{\%Y-\%m-\%d \%H:\%M:\%S}.}
   \item{value}{dendrometer value (\code{µm}).}
   \item{max}{highest measured value up to this timestamp (\code{µm}).}
   \item{twd}{tree water deficit (\code{µm}), i.e. the amount of stem
     shrinkage expressed as the difference between \code{max} and
     \code{value}.}
   \item{gro_yr}{growth since the beginning of the year (\code{µm}). Also
     calculated for years with missing data.}
   \item{frost}{indicates frost periods (i.e. periods in which the
     temperature is below \code{lowtemp}).}
   \item{flags}{character vector specifying the changes that occurred
     during the processing. For more details see the following vignette:
     \href{../doc/Introduction-to-treenetproc.html}{\code{vignette("Introduction-to-treenetproc", package = "treenetproc")}}}
   \item{version}{package version number.}
}
\description{
A dataset containing cleaned (\code{L2}) dendrometer data.
}
\keyword{datasets}
