% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_treenet.R
\name{download_series}
\alias{download_series}
\title{Download Series from Server}
\usage{
download_series(
  meta_series,
  data_format,
  data_version = NULL,
  from,
  to,
  last,
  bind_df,
  reso,
  path_cred,
  server,
  temp_ref,
  tz,
  use_intl
)
}
\arguments{
\item{meta_series}{\code{data.frame}, contains the names of the series to
download including the reference temperature dataset for each series.}

\item{data_format}{character, select processing level of data. Can either
be \code{"L0"} (i.e. raw data), \code{"L1"} (i.e time-aligned) or
\code{"L2"} (i.e. processed).}

\item{data_version}{character, optional argument specifying the package
version number. If provided, \code{L2} data generated by the package
version specified is downloaded from the server (download of \code{L0} or
\code{L1} data is not affected).}

\item{from}{character, optional argument to select data after a specific
date (\code{"YYYY-MM-DD"}).}

\item{to}{character, optional argument to select data up to a specific
date (\code{"YYYY-MM-DD"}).}

\item{last}{numeric, optional argument to specify the number of latest
values. If \code{last} is specified \code{from} and \code{to} need to
be empty.}

\item{bind_df}{logical, indicate whether data should be returned as a
\code{data.frame} (\code{bind_df = TRUE}) or as a named list
\code{bind_df = FALSE}.}

\item{reso}{numeric, desired output time resolution (in minutes). See
\code{Details} for more information on data aggregation.}

\item{path_cred}{optional argument to specify the full path to the
file \code{config.yml} containing the database login data. File can
also be copied to the main folder of the package for automatic
recognition.}

\item{server}{character, specify server from which data is downloaded.
Can be either \code{treenet} or \code{decentlab}.}

\item{temp_ref}{logical, specify whether the reference temperature
dataset(s) should be downloaded along with specified data.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}

\item{use_intl}{logical, specify whether this function is used
internally on the server. Changes \code{stop()} error messages to
\code{message()} only.}
}
\description{
\code{download_series} downloads specified series from TreeNet or
  Decentlab server
}
\keyword{internal,}
\keyword{treenet}
