% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_treenet.R
\name{download_treenet}
\alias{download_treenet}
\title{Download Dendrometer Data from TreeNet or Decentlab Server}
\usage{
download_treenet(site = NULL, sensor_name = NULL, from = NULL,
  to = NULL, server = "treenet", data_format = "L0",
  path_cred = NULL, export = FALSE, bind_df = TRUE, last = NULL,
  tz = "Etc/GMT-1")
}
\arguments{
\item{site}{character, specify the site. String is not case sensitive.}

\item{sensor_name}{character, specify the name of a single or
multiple sensors (e.g. \code{"LWF-Demo-1.dendrometer.ch0"}) of the
same site.}

\item{from}{character, optional argument to select data after a specific
date (\code{"YYYY-MM-DD"}).}

\item{to}{character, optional argument to select data up to a specific
date (\code{"YYYY-MM-DD"}).}

\item{server}{character, specify server from which data is downloaded.
Can be either \code{treenet} or \code{decentlab}.}

\item{data_format}{character, select processing level of data. Can either
be \code{"L0"} (i.e. raw data), \code{"L1"} (i.e time-aligned) or
\code{"L2"} (i.e. processed).}

\item{path_cred}{optional argument to specify the full path to the
file \code{config.yml} containing the database login data. File can
also be copied to the main folder of the package for automatic
recognition.}

\item{export}{logical, indicate whether each sensor will be saved as
an \code{.RData}-file in the current working directory
(\code{export = TRUE}) or will be returned to the console.}

\item{bind_df}{logical, indicate whether data should be returned as a
\code{data.frame} (\code{bind_df = TRUE}) or as a named list
\code{bind_df = FALSE}.}

\item{last}{numeric, optional argument to specify the number of latest
values. If \code{last} is specified \code{from} and \code{to} need to
be empty.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\value{
If \code{export = TRUE} data will be saved in a named list with
  each list element containing one sensor.

  If \code{export = FALSE} data will be saved in the console as a
  \code{data.frame}.
}
\description{
\code{download_treenet} loads \code{L0, L1 or L2} dendrometer data from
  the TreeNet or Decentlab server.
}
\examples{
\dontrun{
download_treenet(site = "lens", export = TRUE, bind_df = FALSE,
                 server = "decentlab")
}
}
\keyword{treenet}
