% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L1.R
\name{fillintergaps}
\alias{fillintergaps}
\title{Interpolate Between Irregular and Regular Timesteps}
\usage{
fillintergaps(df, reso, wnd = reso * 2.1, type = "linear")
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{reso}{desired time resolution of output (in minutes). Resolution
should be chosen to be close to the actual measurement intervals.
\code{reso} needs to be a multiple of 5.}

\item{wnd}{length of time window over which values are interpolated. Defined
as \code{2.1 * reso}. Unit of \code{wnd} is minutes.}

\item{type}{specify type of interpolation between regular timesteps.}
}
\description{
\code{fillintergaps} interpolates gaps between irregular and regular
  timesteps, i.e. only over very small gaps (more an alignement than an
  interpolation).
}
\examples{

}
\keyword{internal}
