% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_corr.R
\name{forcejump}
\alias{forcejump}
\title{Force Jump in Data}
\usage{
forcejump(data_L2, force, n_days = 5)
}
\arguments{
\item{data_L2}{processed dendrometer data as produced by
\code{\link{proc_dendro_L2}}.}

\item{n_days}{numeric, specifies the length of the period (in days) after
the dates specified in \code{force} in which a missed jump is looked for.}
}
\description{
\code{forejump} forces a jump (positive or negative) in the dendrometer
  data that was not corrected during the processing.
}
\keyword{internal}
