% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L1.R
\name{generatets}
\alias{generatets}
\title{Generate Regular Time Stamps}
\usage{
generatets(df, reso, year, tz)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{reso}{desired time resolution of output (in minutes). See
\code{Details} for more information on data aggregation.}

\item{year}{if \code{year = "full"} then the output \code{data.frame}
contains data of complete years, i.e. \code{YYYY-01-01 to YYYY-12-31}; if
\code{year = "asis"} the output \code{data.frame} covers the same period
as the input data.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\description{
\code{generatets} generates regular time stamps over whole period at
  specified resolution (\code{reso}).
}
\keyword{internal}
