% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L1.R
\name{generatets}
\alias{generatets}
\title{Generate Regular Time Stamps}
\usage{
generatets(df, reso, year, tz)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{reso}{numeric, desired output time resolution (in minutes). See
Details for more information on data aggregation.}

\item{year}{character, if \code{year = "asis"} the output
\code{data.frame} covers the same period as the input data;
if \code{year = "full"} then the output \code{data.frame} is expanded to
complete years, i.e. \code{YYYY-01-01 to YYYY-12-31}.}

\item{tz}{specify the desired time zone.}
}
\description{
\code{generatets} generates regular time stamps over whole period at
  specified resolution (\code{reso}).
}
\keyword{internal}
