% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_treenet.R
\name{get_decentlab}
\alias{get_decentlab}
\title{Generate Query for Data from Decentlab Server}
\usage{
get_decentlab(meta_single, path_cred = NULL, tz)
}
\arguments{
\item{meta_single}{character, name of the sensor to download.}

\item{path_cred}{character, path to the file with the credentials.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\description{
\code{get_decentlab} generates a query to download a specified
  series from the decentlab servers.
}
\keyword{internal,}
\keyword{treenet}
