% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_stats.R
\name{phase_stats}
\alias{phase_stats}
\title{Calculate Phase Statistics}
\usage{
phase_stats(
  dendro_L2,
  phase_wnd = 8,
  plot_phase = FALSE,
  plot_export = TRUE,
  agg_daily = TRUE,
  tz = "UTC"
)
}
\arguments{
\item{dendro_L2}{input \code{data.frame} containing cleaned \code{L2}
dendrometer data (obtained from functions \code{\link{proc_dendro_L2}}
or \code{\link{corr_dendro_L2}}).}

\item{phase_wnd}{numeric, specify the window length (in hours) used to
identify local maxima and minima. A shorter window length leads to the
identification of more local maxima and minima.}

\item{plot_phase}{logical, specify whether maxima and minima used for the
calculation of the phase statistics should be plotted.}

\item{plot_export}{logical, specifies whether the plots are exported as a
\code{PDF} file to the working directory or are plotted in the console.}

\item{agg_daily}{logical, specify whether phase statistics are appended
to the \code{L2} data or are exported as a daily aggregated
\code{data.frame}.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\value{
The following additional variables are returned by
  \code{phase_stats}:
   \item{shrink_start/exp_start}{timestamp of the start of the shrinkage
     or expansion phase.}
   \item{shrink_end/exp_end}{timestamp of the end of the shrinkage or
     expansion phase.}
   \item{shrink_dur/exp_dur}{duration (in minutes) of the shrinkage or
     expansion phase.}
   \item{shrink_amp/exp_amp}{amplitude of the shrinkage or expansion phase.}
   \item{shrink_slope/exp_slope}{slope of the shrinkage or expansion phase.}
   \item{phase_class}{days are classified into days on which a single
     shrinkage occurrs during the day \code{(1)}, i.e. where stem radius
     change is likely driven by transpiration; and days on which a single
     expansion occurs \code{(-1)}, i.e. where the stem radius change is
     likely driven by temperature.}

  In case data is not aggregated to daily values
  (\code{agg_daily = FALSE}), all columns are appended to \code{dendro_L2}.
  All parameters related to shrinkage or expansion phases are pasted at
  the timestamp corresponding to the end of the respective phases
  (\code{shrink_end} or \code{exp_end}). All other values are set to
  \code{NA}.
}
\description{
\code{phase_stats} calculates different statistics of shrinkage and
  expansion phases. To identify the phases local maxima and minima are
  identified using two overlapping sets of time windows.
}
\details{
The identification of local maxima and minima in the function
  \code{phase_stats} is inspired by the function
  \code{\link[dendrometeR]{phase_def}} in the package \code{dendrometeR}.
  Overlapping sets of time windows are used to identify local maxima and
  minima, i.e. maxima or minima that appear in both overlapping time windows
  and are not only bound to the start or the end of the respective time
  windows.
}
\examples{

}
