% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_command}
\alias{plot_command}
\title{Plot Command to Plot Dendrometer Data}
\usage{
plot_command(data_L1, data_L2, diff, ...)
}
\arguments{
\item{data_L1}{time-aligned dendrometer data as produced by \code{proc_L1}.}

\item{data_L2}{processed dendrometer data as produced by
\code{proc_dendro_L2}.}

\item{...}{further parameters.}
}
\description{
\code{plot_command()} contains the code necessary to plot dendormeter data.
}
\examples{

}
\keyword{internal}
