% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_frost_period}
\alias{plot_frost_period}
\title{Plot Frost Period}
\usage{
plot_frost_period(data_L2)
}
\arguments{
\item{data_L2}{processed dendrometer data as produced by
\code{\link{proc_dendro_L2}}.}
}
\description{
\code{plot_frost_period} draws a horizontal line in periods of possible
  frost, i.e. when the temperature < \code{lowtemp}.
  This function is exported for its use in vignettes only.
}
\keyword{internal}
