% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_gro_yr_print_vars}
\alias{plot_gro_yr_print_vars}
\title{Plot Yearly Growth Curves and Print Variables}
\usage{
plot_gro_yr_print_vars(data_plot, print_vars, tz)
}
\arguments{
\item{print_vars}{logical, specifies whether used variables should be
plotted in the first panel. Command is only used internally in the
function \code{\link{proc_dendro_L2}}.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\description{
\code{plot_gro_yr_print_vars} plots yearly growth curves for the whole
  period in a single plot and prints the used input variables for plotting.
  In addition, median, maximum and minimum growth values for different
  periods are printed.
}
\keyword{internal}
