% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_interpol_points}
\alias{plot_interpol_points}
\title{Plot Interpolated Points}
\usage{
plot_interpol_points(data_L2)
}
\arguments{
\item{data_L2}{processed dendrometer data as produced by
\code{\link{proc_dendro_L2}}.}
}
\description{
\code{plot_interpol_points} adds points on top of line graph to show
  points that were interpolated.
}
\keyword{internal}
