% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_L1.R
\name{plot_L1}
\alias{plot_L1}
\title{Plot L1 Data}
\usage{
plot_L1(data_L1, plot_period, tz = "Etc/GMT-1", data_L1_orig = NULL,
  plot_name = "L1_plot")
}
\arguments{
\item{data_L1}{time-aligned dendrometer data as produced by
\code{\link{proc_L1}}.}

\item{plot_period}{specify whether plots should be displayed over the whole
plot_period (\code{plot_period = "full"}), for each year separately
(\code{plot_period = "yearly"}) or for each month
(\code{plot_period = "monthly"}).}

\item{tz}{specify the desired time zone. Default is \code{"Etc/GMT-1"}.}

\item{data_L1_orig}{optional, used in \code{\link{corr_dendro_L1}}. Can be
used to plot a previous version of \code{data_L1} in the background.}

\item{plot_name}{character, specify name of the PDF in which the plots are
saved.}
}
\value{
Plots are saved to current working directory as
  \code{L1_plot.pdf} or as specified in \code{plot_name}.
}
\description{
\code{plot_L1} plots \code{L1} data divided into specified periods. Can be
  used for both dendrometer and climate data.
}
\examples{

}
