% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_dendro_L2.R
\name{proc_dendro_L2}
\alias{proc_dendro_L2}
\title{Process L1 Dendrometer Data to L2}
\usage{
proc_dendro_L2(dendro_data, temp_data = NULL, tol_out = 10,
  tol_jump = 50, lowtemp = 5, frost_thr = 5, interpol = NULL,
  frag_len = NULL, plot = TRUE, plot_period = "full",
  plot_show = "all", plot_export = TRUE, plot_name = "proc_L2_plot",
  plot_mds = FALSE, iter_clean = 1, tz = "UTC")
}
\arguments{
\item{dendro_data}{\code{data.frame} with time-aligned dendrometer
data. Output of function \code{proc_L1()}.}

\item{temp_data}{\code{data.frame} with time-aligned temperature data.
Output of function \code{proc_L1()} (see Details for further information).}

\item{tol_out}{numeric, defines the tolerance of the threshold above or
below which a value is classified as outlier (see Details for further
information).}

\item{tol_jump}{numeric, defines the tolerance of the threshold above or
below which a value is flagged for jump correction.}

\item{lowtemp}{numeric, specifies temperature in °C below which shrinkage
in stem diameter due to frost is expected. Default value is set to
\code{5°C} due to hysteresis shortly before or after frost events.}

\item{frost_thr}{numeric, increases the thresholds for outlier
classifiation in periods of probable frost (i.e. temperature <
\code{lowtemp}). The thresholds are multiplied by the value provided.}

\item{interpol}{numeric, length of gaps (in minutes) in which values are
linearly interpolated. Set \code{interpol = 0} to disable gapfilling.
If \code{interpol = NULL} the default value is set to
\code{interpol = 2.1 * reso}.}

\item{frag_len}{numeric, specifies the length of data fragments in-between
\code{NA} data that are automatically deleted during data cleaning. This
can be helpful to remove short fragments of erroneous data within a
period of missing data, i.e. after jumps. If \code{frag_len = NULL} the
devault value is set to \code{frag_len = 2.1}.}

\item{plot}{logical, specify whether a comparison of \code{L1} and \code{L2}
data should be plotted.}

\item{plot_period}{specify whether plots should be displayed over the whole
plot_period (\code{plot_period = "full"}), for each year separately
(\code{plot_period = "yearly"}) or for each month
(\code{plot_period = "monthly"}).}

\item{plot_show}{specify whether all periods should be plotted
(\code{plot_show = "all"}) or only those periods in which \code{L1} and
\code{L2} dendrometer data differ after processing
(\code{plot_show = "diff"}).}

\item{plot_export}{logical, specifies whether the plots are exported as a
\code{PDF} file to the working directory or are plotted in the console.}

\item{plot_name}{character, specify name of the PDF in which the plots are
saved.}

\item{plot_mds}{logical, specify whether maxima and minima used for the
calculation of the maximum daily shrinkage (\code{mds}) should be
plotted.}

\item{iter_clean}{numeric, specifies the number of times the cleaning
process is repeated. Can be used to check whether running the cleaning
process many times has an effect on the results.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\value{
The function returns a \code{data.frame} with processed dendrometer
 data containing the following columns:
   \item{series}{name of the series.}
   \item{ts}{timestamp with format \code{\%Y-\%m-\%d \%H:\%M:\%S}.}
   \item{value}{dendrometer value.}
   \item{max}{highest measured value up to this timestamp.}
   \item{twd}{tree water deficit, i.e. the amount of stem shrinkage
     expressed as the difference between \code{max} and \code{value}.}
   \item{mds}{maximum daily shrinkage, calculated as the difference between
     a local maximum that occurs before a local minimum during one day. If
     there is no local maximum or minimum or if the minimum occurs prior to
     the maximum, then \code{mds = NA}. This may occur on days with rain or
     in winter (see Details for further information).}
   \item{gro_yr}{growth since the beginning of the year. Also calculated
     for years with missing data.}
   \item{gro_start}{day of year at which growth starts. \code{gro_start} is
     defined as the day of year at which 5\% of total yearly growth is
     surpassed.}
   \item{gro_end}{day of year at which growth stops. \code{gro_end} is
     defined as the day of year at which 95\% of total yearly growth is
     reached.}
   \item{flags}{character vector specifying the changes that occurred
     during the processing. For more details see the following vignette:
     \href{../doc/Introduction-to-treenetproc.html}{\code{vignette("Introduction-to-treenetproc", package = "treenetproc")}}}
   \item{version}{package version that was used.}
}
\description{
\code{proc_dendro_L2} processes time-aligned (\code{L1}) dendrometer data
  to processed (\code{L2}) dendrometer data. It removes jumps, corrects
  small gaps and calculates growth, tree water deficit and maximum daily
  shrinkage.
}
\details{
\code{temp_data} is used to define periods in which frost shrinkage
  is probable, e.g. when the temperature is <5°C. Without temperature data,
  shrinkages due to frost may be classified as outliers.

  \code{temp_data} can also be attached to dendrometer data. In this case,
  the \code{series} name of temperature data has to contain the string
  \code{temp}. In case no temperature dataset is specified, a sample
  temperature dataset will be used with a warning. The sample temperature
  dataset assumes probable frost shringkage in the months December, January
  and February.

  Outliers are classified based on the value difference (referred to as
  \code{diff}) between two timesteps. A \code{diff} is classified as an
  outlier if it deviates more than a threshold value from the first
  or third quartile of all \code{diff}. Thresholds are calculated as:\cr
  \code{threshold_low = quantile(diff, probs = 0.25) + tol *
  \link[stats]{mad}(diff)}\cr
  \code{threshold_high = quantile(diff, probs = 0.75) + tol *
  \link[stats]{mad}(diff)}

  Thus, \code{tol} describes the number of times \code{\link[stats]{mad}}
  is added to the first or third quartile of \code{diff}, before a
  \code{diff} is classified as an outlier. \code{tol} is increased to
  \code{tol * frost_thr} in periods of probable frost (i.e. in periods
  where the air temperature is below \code{lowtemp}).

  The maximum daily shrinkage \code{mds} is calculated similarly as in
  the function \code{\link[dendrometeR]{phase_def}} in the package
  \code{dendrometeR}. First, local maxima and minima are identified
  using a moving window. \code{mds} is only calculated if a local maximum
  occurs before a local minimum (i.e. if the stem shrinks during the day).
}
\examples{
proc_dendro_L2(dendro_data = dendro_data_L1, plot_period = "monthly",
               plot_export = FALSE)

}
