% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_dendro_L2.R
\name{proc_dendro_L2}
\alias{proc_dendro_L2}
\title{Clean Time-aligned Dendrometer Data}
\usage{
proc_dendro_L2(
  dendro_L1,
  temp_L1 = NULL,
  tol_out = 10,
  tol_jump = 50,
  lowtemp = 5,
  frost_thr = 5,
  interpol = NULL,
  frag_len = NULL,
  plot = TRUE,
  plot_period = "full",
  plot_show = "all",
  plot_export = TRUE,
  plot_name = "proc_L2_plot",
  iter_clean = 1,
  tz = "UTC"
)
}
\arguments{
\item{dendro_L1}{\code{data.frame} with time-aligned dendrometer
data. Output of function \code{\link{proc_L1}}.}

\item{temp_L1}{\code{data.frame} with time-aligned temperature data.
Output of function \code{\link{proc_L1}} (see Details for further
information).}

\item{tol_out}{numeric, defines the rigidity of the threshold above or
below which a value is classified as an outlier. Lower values
increase the rigidity (see Details for further information).}

\item{tol_jump}{numeric, defines the rigidity of the threshold above or
below which a value is flagged for jump correction. Lower values
increase the rigidity (see Details for further information).}

\item{lowtemp}{numeric, specifies the temperature in 째C below which frost
shrinkage or expansion is expected. Default value is set to
\code{5째C} due to hysteresis shortly before or after frost events.}

\item{frost_thr}{numeric, increases the thresholds for outlier
and jump detection in periods of probable frost (i.e. temperature <
\code{lowtemp}). The thresholds are multiplied by the value provided.}

\item{interpol}{numeric, length of gaps (in minutes) for which values are
linearly interpolated after data cleaning. Set \code{interpol = 0} to
disable gapfilling. If \code{interpol = NULL} the default value is set to
\code{interpol = 2.1 * reso} (i.e. two timestamps).}

\item{frag_len}{numeric, specifies the length of data fragments occurring
in-between missing data that are automatically deleted during data
cleaning. This can be helpful to remove short fragments of erroneous data
within a period of missing data, i.e. after jumps. If
\code{frag_len = NULL} the default value is set to \code{frag_len = 2.1}.}

\item{plot}{logical, specify whether the changes that occurred during data
cleaning should be plotted.}

\item{plot_period}{character, specify whether plots should be displayed for
the whole period (\code{plot_period = "full"}), or separately for each
year (\code{plot_period = "yearly"}) or for each month
(\code{plot_period = "monthly"}).}

\item{plot_show}{character, specify whether all periods should be plotted
(\code{plot_show = "all"}) or only those periods in which \code{L1}
dendrometer data changed during data cleaning (\code{plot_show = "diff"}).}

\item{plot_export}{logical, specifies whether the plots are exported as a
\code{PDF} file to the working directory or are plotted in the console.}

\item{plot_name}{character, specify name of the PDF to which the plots are
saved.}

\item{iter_clean}{numeric, specifies the number of times the cleaning
process is repeated. Can be used to check whether running the cleaning
process multiple times has an effect on the results. In most cases, a
single iteration is sufficient.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\value{
The function returns a \code{data.frame} with processed dendrometer
  data containing the following columns:
   \item{series}{name of the dendrometer series.}
   \item{ts}{timestamp with format \code{\%Y-\%m-\%d \%H:\%M:\%S}.}
   \item{value}{dendrometer value.}
   \item{max}{highest measured value up to this timestamp.}
   \item{twd}{tree water deficit, i.e. the amount of stem shrinkage
     expressed as the difference between \code{max} and \code{value}.}
   \item{gro_yr}{growth since the beginning of the year. Also calculated
     for years with missing data.}
   \item{frost}{indicates frost periods (i.e. periods in which the
     temperature is below \code{lowtemp}).}
   \item{flags}{character vector specifying the changes that occurred
     during the processing. For more details see the following vignette:
     \href{../doc/Introduction-to-treenetproc.html}{\code{vignette("Introduction-to-treenetproc", package = "treenetproc")}}}
   \item{version}{package version number.}
}
\description{
\code{proc_dendro_L2} cleans time-aligned (\code{L1}) dendrometer data
  by removing outliers and correcting for erroneous jumps or shifts in
  the data.
}
\details{
Time-aligned temperature data \code{temp_L1} is used to define
  periods in which frost shrinkage is probable, e.g. when the temperature
  is below \code{lowtemp}. Without temperature data, frost shrinkages may be
  classified as outliers. For more details and an example see the following
  vignette:
  \href{../doc/Introduction-to-treenetproc.html}{\code{vignette("Introduction-to-treenetproc", package = "treenetproc")}}.

  Temperature data can also be provided along with dendrometer data. In this
  case, the name of the temperature series has to contain the string
  \code{"temp"}. In case no temperature dataset is specified, a sample
  temperature dataset will be used with a warning. The sample temperature
  dataset assigns permanent frost to the three months December, January
  and February.

  Outliers and jumps are identified when exceeding a lower or upper
  threshold. Thresholds are obtained on the basis of density distributions
  of differences between neighbouring data points. The rigidity of the
  thresholds can be controlled with the arguments \code{tol_jump} and
  \code{tol_out}. For more information on the calculation of the thresholds
  the user is referred to Kn체sel et al. (2020, in prep).
}
\examples{
proc_dendro_L2(dendro_L1 = dendro_data_L1, plot_period = "monthly",
               plot_export = FALSE)

}
\references{
Kn체sel S., Haeni M., Wilhelm M., Peters R.L., Zweifel R. 2020.
  treenetproc - An R package to clean, process and visualise dendrometer
  data. In preparation.
}
