% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_dendro_L2.R
\name{proc_dendro_L2}
\alias{proc_dendro_L2}
\title{Process L1 Dendrometer Data to L2}
\usage{
proc_dendro_L2(dendro_data, temp_data = NULL, val_range = c(0, 20000),
  diffwin = 2000, diffsum = 1000, lowtemp = 5, tz = "Etc/GMT-1")
}
\arguments{
\item{dendro_data}{\code{data.frame} with time-aligned dendrometer
data. Output of function \code{proc_L1}.}

\item{temp_data}{\code{data.frame} with time-aligned temperature data.
Output of function \code{proc_L1} (see Details for further information).}

\item{val_range}{numeric vector specifying the minimum and maximum
\code{c(min, max)} of credible dendrometer measurement values. Values
lower than \code{min} or higher than \code{max} are deleted without
notice.}

\item{diffwin}{maximal hourly difference expected in winter.}

\item{diffsum}{maximal hourly difference expected in summer.}

\item{lowtemp}{numeric, specifies temperature in °C below which shrinkage
in stem diameter due to frost is expected.}

\item{tz}{specify the desired time zone. Default is \code{"Etc/GMT-1"}.}
}
\value{
The function returns:
 a \code{data.frame} with processed dendrometer data containing the
 following columns
   \item{series}{name of the series.}
   \item{ts}{timestamp with format \code{\%Y-\%m-\%d \%H:\%M:\%S}.}
   \item{value}{dendrometer value.}
   \item{max}{highest measured value up to this timestamp.}
   \item{twd}{tree water deficit, i.e. the amount of stem shrinkage
     expressed as the difference between \code{max} and \code{value}.}
   \item{mds}{maximum daily shrinkage, calculated as the difference between
     a local maximum that occurs before a local minimum during one day. If
     there is no local maximum or minimum or if the minimum occurs prior to
     the maximum, then \code{mds = NA}. This may occur on days with rain or
     in winter.}
   \item{gro_yr}{growth since the beginning of the year. Also calculated if
     for years with missing data.}
   \item{flags}{number specifying whether and which changes occurred during
     the processing.}
   \item{version}{processing version.}
}
\description{
\code{proc_dendro_L2()} processes time-aligned (\code{L1}) dendrometer data
  to processed (\code{L2}) dendrometer data. It removes jumps, corrects
  small gaps and calculates growth, tree water deficit and maximum daily
  shrinkage.
}
\details{
\code{temp_data} is used to define periods in which frost shrinkage
  is probable, i.e. when temperature is <5°C. Without temperature data,
  frost shrinkage may be classified as outliers.

  \code{temp_data} can also contain other climate data. In this case, the
  \code{series} name of temperature data has to contain the string
  \code{temp}. In case no temperature dataset is specified, a sample
  temperature dataset will be used with a warning. The sample temperature
  dataset assumes probable frost shringkage in the months December, January
  and February.
}
\examples{
\dontrun{
proc_dendro_L2(dendro_data = data_L1_dendro, temp_data = data_L1_temp,
               val_range = c(0, 20000))
}

}
