% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_L1.R
\name{proc_L1}
\alias{proc_L1}
\title{Process Raw Dendrometer or Meteorological Data to L1}
\usage{
proc_L1(data, reso = 10, year = "asis", input = "long",
  date_format = "\%Y-\%m-\%d \%H:\%M:\%S", tz = "UTC")
}
\arguments{
\item{data}{input \code{data.frame} containing raw dendrometer or
meteorological data (see Details for formatting requirements).}

\item{reso}{desired time resolution of output (in minutes). See
\code{Details} for more information on data aggregation.}

\item{year}{if \code{year = "full"} then the output \code{data.frame}
contains data of complete years, i.e. \code{YYYY-01-01 to YYYY-12-31}; if
\code{year = "asis"} the output \code{data.frame} covers the same period
as the input data.}

\item{input}{specify the way the input \code{data.frame} is structured. Can
be either in \code{"long"} format (i.e. sensors below each other with a
\code{series}) or in \code{"wide"} format (i.e. one sensor per column).}

\item{date_format}{specify a custom date format if it is not
\code{\%Y-\%m-\%d \%H:\%M:\%S}.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\value{
a \code{data.frame} with measurements aligned to regular time
  intervals (interval specified in \code{reso}) containing the following
  columns:
   \item{series}{name of the series.}
   \item{ts}{timestamp with format \code{\%Y-\%m-\%d \%H:\%M:\%S}.}
   \item{value}{dendrometer value.}
   \item{version}{package version that was used.}
}
\description{
\code{proc_L1} processes raw dendrometer and raw meteorological data to
  time-aligned data.
}
\details{
The input data needs to contain a timestamp column
  (named \code{ts}). If \code{ts} is not provided in the standard format
  (\code{\%Y-\%m-\%d \%H:\%M:\%S}) the format needs to be specified in
  \code{format_date}.\cr
  Data of multiple sensors can either be in \code{long}
  format (\code{input = "long"}) with a column named \code{series} to
  differentiate the sensors or in \code{wide} format (\code{input = "wide"})
  with sensors in separate columns.\cr
  Temperature and precipitation data must contain \code{temp} or
  \code{prec}, respectively, somewhere in their series name.

  To align data at regular time intervals defined in \code{reso}, the data
  is linearly interpolated between the two values measured closest to the
  regular time intervals. The linear interpolation is restricted to gaps
  that are smaller than \code{2.1 * reso}. If data is already provided at
  regular time steps, the corresponding values are selected. For more
  information see the following vignette: For more details see the
  following vignette:
  \href{../doc/Functionality-of-treenetproc.html}{\code{vignette("Functionality-of-treenetproc", package = "treenetproc")}}
}
\examples{
proc_L1(data = dendro_data_L0)
proc_L1(data = dendro_data_L0_wide, input = "wide")

}
