% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_treenet.R
\name{proc_treenet}
\alias{proc_treenet}
\title{Process Treenet Data to L1 or L2}
\usage{
proc_treenet(site = NULL, sensor_name = NULL, from = NULL,
  to = NULL, temp_name = NULL, reso = 10, year = "asis", tol = 9,
  iter_clean = 3, interpol = 120, lowtemp = 5, tz = "Etc/GMT-1",
  version = "L2", plot = TRUE, plot_period = "full",
  plot_export = TRUE, plot_name = "proc_L2_plot", plot_show = "all",
  plot_mds = FALSE, path_cred = NULL)
}
\arguments{
\item{site}{character, specify the site. String is not case sensitive.}

\item{sensor_name}{character, specify the name of a single or
multiple sensors (e.g. \code{"LWF-Demo-1.dendrometer.ch0"}) of the
same site.}

\item{from}{character, optional argument to select data after a specific
date (\code{"YYYY-MM-DD"}).}

\item{to}{character, optional argument to select data up to a specific
date (\code{"YYYY-MM-DD"}).}

\item{temp_name}{character, optional argument to specify the series name
of a temperature dataset. Needed if there are no or multiple on-site air
temperature measurements. If \code{temp_name} is not specified and no
temperature dataset is found on the server, a sample dataset will be
used with a warning.}

\item{reso}{desired time resolution of output (in minutes). See
\code{Details} for more information on data aggregation.}

\item{year}{if \code{year = "full"} then the output \code{data.frame}
contains data of complete years, i.e. \code{YYYY-01-01 to YYYY-12-31}; if
\code{year = "asis"} the output \code{data.frame} covers the same period
as the input data.}

\item{tol}{numeric, defines the tolerance of the threshold values above or
below which dendrometer measurements are classified as outliers (see
Details for further information).}

\item{iter_clean}{numeric, specifies the number of times the cleaning
process is repeated. Increase if jumps are not corrected after
processing.}

\item{interpol}{numeric, length of gaps (in minutes) in which values are
linearly interpolated.}

\item{lowtemp}{numeric, specifies temperature in °C below which shrinkage
in stem diameter due to frost is expected. Default value is set to
\code{5°C} due to hysteresis shortly before or after frost events.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}

\item{version}{character, specify to which level the data should be
processed. Can be one of \code{L0, L1} or \code{L2}.}

\item{plot}{logical, specify whether a comparison of \code{L1} and \code{L2}
data should be plotted.}

\item{plot_period}{specify whether plots should be displayed over the whole
plot_period (\code{plot_period = "full"}), for each year separately
(\code{plot_period = "yearly"}) or for each month
(\code{plot_period = "monthly"}).}

\item{plot_export}{logical, specifies whether the plots are exported as a
\code{PDF} file to the working directory or are plotted in the console.}

\item{plot_name}{character, specify name of the PDF in which the plots are
saved.}

\item{plot_show}{specify whether all periods should be plotted
(\code{plot_show = "all"}) or only those periods in which \code{L1} and
\code{L2} dendrometer data differ after processing
(\code{plot_show = "diff"}).}

\item{plot_mds}{logical, specify whether maxima and minima used for the
calculation of the maximum daily shrinkage (\code{mds}) should be
plotted.}

\item{path_cred}{optional argument to specify the full path to the
file \code{config.yml} containing the database login data. File can
also be copied to the main folder of the package for automatic
recognition.}
}
\value{
The function returns a \code{data.frame} with processed dendrometer
 data containing the following columns:
   \item{series}{name of the series.}
   \item{ts}{timestamp with format \code{\%Y-\%m-\%d \%H:\%M:\%S}.}
   \item{value}{dendrometer value.}
   \item{max}{highest measured value up to this timestamp.}
   \item{twd}{tree water deficit, i.e. the amount of stem shrinkage
     expressed as the difference between \code{max} and \code{value}.}
   \item{mds}{maximum daily shrinkage, calculated as the difference between
     a local maximum that occurs before a local minimum during one day. If
     there is no local maximum or minimum or if the minimum occurs prior to
     the maximum, then \code{mds = NA}. This may occur on days with rain or
     in winter (see Details for further information).}
   \item{gro_yr}{growth since the beginning of the year. Also calculated
     for years with missing data.}
   \item{gro_start}{day of year at which growth starts. \code{gro_start} is
     defined as the day of year at which 5\% of total yearly growth is
     surpassed.}
   \item{gro_end}{day of year at which growth stops. \code{gro_end} is
     defined as the day of year at which 95\% of total yearly growth is
     reached.}
   \item{flags}{character vector specifying whether and which changes
     occurred during the processing. For more details see the following
     vignette: \href{../doc/Introduction-to-treenetproc.html}{\code{vignette("Introduction-to-treenetproc", package = "treenetproc")}}}
   \item{version}{processing version.}
}
\description{
\code{proc_treenet} processes \code{L0} dendrometer data from the
  treenet server directly to \code{L1} (i.e. time-aligned data) or
  \code{L2} (i.e. cleaned \code{L1} data).
}
\details{
\code{temp_data} is used to define periods in which frost shrinkage
  is probable, i.e. when temperature is <5°C. Without temperature data,
  frost shrinkage may be classified as outliers.

  \code{temp_data} can also contain other climate data. In this case, the
  \code{series} name of temperature data has to contain the string
  \code{temp}. In case no temperature dataset is specified, a sample
  temperature dataset will be used with a warning. The sample temperature
  dataset assumes probable frost shringkage in the months December, January
  and February.

  Outliers are classified based on the value difference (referred to as
  \code{diff}) between two timesteps. A \code{diff} is classified as an
  outlier if it deviates more than a threshold value from the first
  or third quartile of all \code{diff} in a specific time window.
  Thresholds are calculated as:\cr
  \code{threshold_low = quantile(diff, probs = 0.25) + tol *
  \link[stats]{mad}(diff)}\cr
  \code{threshold_high = quantile(diff, probs = 0.75) + tol *
  \link[stats]{mad}(diff)}

  Thus, \code{tol} describes the number of times \code{\link[stats]{mad}}
  is added to the first or third quartile of \code{diff} in a specific time
  window until \code{diff} is classified as an outlier. \code{tol} is
  increased to \code{tol * 15} in periods of probable frost (i.e. in
  periods where the air temperature is below \code{lowtemp}).

  The maximum daily shrinkage \code{mds} is calculated similarly as in
  the function \code{\link[dendrometeR]{phase_def}} in the package
  \code{dendrometeR}. First, local maxima and minima are identified
  using a moving window. \code{mds} is only calculated if a local maximum
  occurs before a local minimum (i.e. if the stem shrinks during the day).
}
\examples{
\dontrun{
proc_treenet(site = "sagno", year = "full", plot_period = "yearly")

proc_treenet(site = "lens", plot_period = "monthly")

proc_treenet(sensor_name = c("Alvaneu-2.dendrometer.ch0",
                             "Alvaneu-4.dendrometer.ch0"))

# throws error
proc_treenet(sensor_name = c("Alvaneu-2.dendrometer.ch0",
                             "Alvaneu-4.dendrometer.ch0",
                             "Bachtel-2.dendrometer.ch0"))
}
}
\seealso{
\code{\link{proc_L1}} to process dendrometer or climate data to
  \code{L1}.\cr
  \code{\link{proc_dendro_L2}} to process dendrometer data from \code{L1}
  to \code{L2}.\cr
  \code{\link{corr_dendro_L1}} to correct errors before processing.\cr
  \code{\link{corr_dendro_L3}} to correct errors after processing.
}
\keyword{treenet}
