% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_treenet_L2.R
\name{proc_treenet_L2}
\alias{proc_treenet_L2}
\title{Process Treenet Data to L2}
\usage{
proc_treenet_L2(site = NULL, sensor_name = NULL, from = NULL,
  to = NULL, path_cred = NULL, temp_name = NULL, reso = 10,
  year = "asis", n_mad = 9, iter_clean = 3, lowtemp = 5,
  tz = "Etc/GMT-1", plot = TRUE, plot_period = "full",
  plot_name = "proc_L2_plot", plot_show = "all", plot_mds = FALSE)
}
\arguments{
\item{site}{character string specifying the site. String is not
case sensitive.}

\item{sensor_name}{character string specifying the name of a single or
multiple sensors (e.g. \code{"LWF-Demo-1.dendrometer.ch0"}) of the
same site.}

\item{from}{optional argument to select data after a specific date
(\code{"YYYY-MM-DD"}).}

\item{to}{Optional argument to select data up to a specific date
(\code{"YYYY-MM-DD"}).}

\item{path_cred}{optional argument to specify the full path to the
file \code{config.yml} containing the database login data. File can
also be copied to the main folder of the package for automatic
recognition.}

\item{temp_name}{optional argument to specify the series name of a
temperature dataset. Needed if there are no or multiple on-site air
temperature measurements. If \code{temp_name} is not specified and no
temperature dataset is found on the server, a sample dataset will be
used with a warning.}

\item{reso}{desired time resolution of output (in minutes). Resolution
should be chosen to be close to the actual measurement intervals.
\code{reso} needs to be a multiple of 5.}

\item{year}{if \code{year = "full"} then the output \code{data.frame}
contains data of complete years, i.e. \code{YYYY-01-01 to YYYY-12-31}; if
\code{year = "asis"} the output \code{data.frame} covers the same period
as the input data.}

\item{n_mad}{numeric, defines the thresholds above or below which values
are classified as outliers (see Details for further information).}

\item{iter_clean}{numeric, specifies the number of times the cleaning
process is repeated. Increase if jumps are not corrected after
processing.}

\item{lowtemp}{numeric, specifies temperature in °C below which shrinkage
in stem diameter due to frost is expected. Default value is set to
\code{5°C} due to hysteresis shortly before or after frost events.}

\item{tz}{specify the desired time zone. Default is \code{"Etc/GMT-1"}.}

\item{plot}{logical, specify whether a comparison of \code{L1} and \code{L2}
data should be plotted.}

\item{plot_period}{specify whether plots should be displayed over the whole
plot_period (\code{plot_period = "full"}), for each year separately
(\code{plot_period = "yearly"}) or for each month
(\code{plot_period = "monthly"}).}

\item{plot_name}{character, specify name of the PDF in which the plots are
saved.}

\item{plot_show}{specify whether all periods should be plotted
(\code{plot_show = "all"}) or only those periods in which \code{L1} and
\code{L2} dendrometer data differ after processing
(\code{plot_show = "diff"}).}

\item{plot_mds}{logical, specify whether maxima and minima used for the
calculation of the maximum daily shrinkage (\code{mds}) should be
plotted.}
}
\value{
The function returns:
 a \code{data.frame} with processed dendrometer data containing the
 following columns
   \item{series}{name of the series.}
   \item{ts}{timestamp with format \code{\%Y-\%m-\%d \%H:\%M:\%S}.}
   \item{value}{dendrometer value.}
   \item{max}{highest measured value up to this timestamp.}
   \item{twd}{tree water deficit, i.e. the amount of stem shrinkage
     expressed as the difference between \code{max} and \code{value}.}
   \item{mds}{maximum daily shrinkage, calculated as the difference between
     a local maximum that occurs before a local minimum during one day. If
     there is no local maximum or minimum or if the minimum occurs prior to
     the maximum, then \code{mds = NA}. This may occur on days with rain or
     in winter.}
   \item{gro_yr}{growth since the beginning of the year. Also calculated if
     for years with missing data.}
   \item{flags}{number specifying whether and which changes occurred during
     the processing.}
   \item{version}{processing version.}
}
\description{
\code{proc_treenet_L2()} processes \code{L0} dendrometer data from the
  treenet server directly to \code{L2}.
}
\examples{
\dontrun{
proc_treenet_L2(site = "sagno", year = "full", plot_period = "yearly")

proc_treenet_L2(site = "lens", plot_period = "monthly")

proc_treenet_L2(sensor_name = c("Alvaneu-2.dendrometer.ch0",
                "Alvaneu-4.dendrometer.ch0"))

# throws error
proc_treenet_L2(sensor_name = c("Alvaneu-2.dendrometer.ch0",
                "Alvaneu-4.dendrometer.ch0", "Bachtel-2.dendrometer.ch0"))
}
}
\seealso{
\code{\link{proc_L1}} to process dendrometer or climate data to
  \code{L1} only (i.e. time-aligned data), \code{\link{proc_dendro_L2}}
  to process dendrometer data from \code{L1} to \code{L2}.
  \code{\link{corr_dendro_L3}} to correct errors in processing.
}
