% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_phase.R
\name{removeconsec}
\alias{removeconsec}
\title{Remove Consecutive Maxima or Minima}
\usage{
removeconsec(df, remove, notremove, mode)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{remove}{numeric, for which consecutive values are removed
(i.e. maxima or minima).}

\item{notremove}{numeric, for which consecutive values are not removed
(i.e. maxima or minima).}

\item{mode}{character, specifies whether consecutive maxima or minima
are removed. Can either be \code{"max"} or \code{"min"}.}
}
\description{
\code{removeconsec} removes consecutive maxima or minima and keeps only
  the higher or lower value, respectively. The function makes sure no two
  maxima or minima appear consecutively. It is a helper function of
  \code{\link{phase_stats}}.
}
\keyword{internal}
