% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_corr.R
\name{removecorr}
\alias{removecorr}
\title{Removes Wrong Corrections in Data}
\usage{
removecorr(data_L1, data_L2, remove, tz)
}
\arguments{
\item{data_L1}{time-aligned dendrometer data as produced by
\code{\link{proc_L1}}.}

\item{data_L2}{processed dendrometer data as produced by
\code{\link{proc_dendro_L2}}.}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\description{
\code{removecorr} removes wrong corrections in the dendrometer data.
}
\keyword{internal}
