% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L3.R
\name{removejump}
\alias{removejump}
\title{Removes Wrong Jump in Data}
\usage{
removejump(data_L1, data_L2, remove, tz)
}
\arguments{
\item{data_L1}{time-aligned dendrometer data as produced by
\code{\link{proc_L1}}. Optional, only needed for \code{remove} and if
\code{plot = TRUE}.}

\item{data_L2}{processed dendrometer data as produced by
\code{\link{proc_dendro_L2}}.}

\item{remove}{numeric vector, specify numbers of the changes that should
be removed. Numbers are displayed in the plots produced by
\code{\link{proc_dendro_L2}} or \code{\link{plot_proc_L2}} (numbers
are only displayed if \code{plot_period = "monthly"}).}

\item{tz}{specify the desired time zone. Default is \code{"UTC"}.}
}
\description{
\code{removejump} removes wrong corrections in the dendrometer data.
}
\keyword{internal}
