% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_L1.R
\name{roundtimetoreso}
\alias{roundtimetoreso}
\title{Rounds Irregular Time Stamp to Regular}
\usage{
roundtimetoreso(df, reso, pos, tz)
}
\arguments{
\item{df}{input \code{data.frame}.}

\item{reso}{time resolution between two timestamps (in minutes). Resolution
should be chosen to be close to the actual measurement intervals.
\code{reso} needs to be a multiple of 5.}

\item{tz}{specify the desired time zone. Default is \code{"Etc/GMT-1"}.}
}
\description{
\code{roundtimetoreso} rounds first and last irregular timestep to
  next or previous regular timestep.
}
\examples{

}
\keyword{internal}
