% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_data.R
\name{select_data}
\alias{select_data}
\title{Select Dendrometer and Temperature Data from Treenet Server}
\usage{
select_data(site = NULL, sensor_name = NULL, from = NULL,
  to = NULL, path_cred = NULL, temp_name = NULL)
}
\arguments{
\item{site}{character string specifying the site. String is not
case sensitive.}

\item{sensor_name}{character string specifying the name of a single or
multiple sensors (e.g. \code{"LWF-Demo-1.dendrometer.ch0"}) of the
same site.}

\item{from}{optional argument to select data after a specific date
(\code{"YYYY-MM-DD"}).}

\item{to}{Optional argument to select data up to a specific date
(\code{"YYYY-MM-DD"}).}

\item{path_cred}{optional argument to specify the full path to the
file \code{config.yml} containing the database login data. File can
also be copied to the main folder of the package for automatic
recognition.}

\item{temp_name}{optional argument to specify the series name of a
temperature dataset. Needed if there are no or multiple on-site air
temperature measurements. If \code{temp_name} is not specified and no
temperature dataset is found on the server, a sample dataset will be
used with a warning.}
}
\description{
\code{select_data()} accesses and downloads dendrometer and temperature
  data from treenet server. Function imported and adapted from package
  \code{treenetdown}.
}
\examples{
\dontrun{
select_data(site = "birmensdo")
}

}
\keyword{internal}
