% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_corr.R
\name{summariseflagscorr}
\alias{summariseflagscorr}
\title{Summarise Flags}
\usage{
summariseflagscorr(df, force = NULL, delete = NULL, force.now = NULL)
}
\arguments{
\item{df}{input \code{data.frame}.}
}
\description{
\code{summariseflagscorr} appends the flags of the corrections to the
  existing flags (except for reverse).
}
\keyword{internal}
